refactor!: split settings page into URL-routed tabs
The 3487-line settings/page.tsx is broken into a layout + per-tab pages
under /settings/[tab], each with its own bundle and shareable URL.

Layout
- src/app/(dashboard)/settings/layout.tsx: tab switcher (next/link), saving
  indicator, page header. Wraps children in SettingsProvider.
- src/app/(dashboard)/settings/_lib/SettingsContext.tsx: provider that owns
  shared form state (server name, email, registries, template values),
  nodes + node health, persistSettings, submitNode, and the SSH setup modal.
- src/app/(dashboard)/settings/_lib/helpers.ts: pure helpers + types
  (formatBytes, groupFilesByService, log status badges, restore-overlay
  icon labels, AppUpdateStatus, BackupStreamEvent…)

Routes
- /settings: server-side redirect to /settings/nodes (5 LOC)
- /settings/nodes: node list, add/edit/delete, deploy SSH key, terminal
- /settings/backups: backup list + create/restore/delete, backup sync,
  nginx config export/import, restore overlay, file preview
- /settings/notifications: SMTP form + recipients
- /settings/updates: app update status, channel switch, auto-update toggle,
  update progress modal, confirm modal
- /settings/advanced: server identity, template variables, registries,
  log level

URL routing means each tab is now bookmarkable, sharable, and gets its own
chunk. The breaking surface change (\`/settings\` → \`/settings/nodes\`) is
why this is refactor!

Validation
- All 243 tests pass
- Build clean — registers /settings/{advanced,backups,nodes,notifications,updates}
- Lint + typecheck clean on all new files
