+++ src/app/(dashboard)/settings/page.tsx
@@ -1,3487 +1,5 @@
-'use client';
+import { redirect } from 'next/navigation';
 
-import { useState, useEffect, useCallback, useRef } from 'react';
-import { useRouter } from 'next/navigation';
-import { Save, Mail, Plus, Trash2, RefreshCw, Download, Clock, GitBranch, Loader2, CheckCircle2, XCircle, Server, Key, Terminal, Edit2, ShieldAlert, WifiOff, Globe, HardDrive, RotateCcw, UploadCloud, X, Eye, ChevronDown, ChevronRight, Settings, Activity, FolderOpen, Database, Shield, Upload } from 'lucide-react';
-import { useToast } from '@/providers/ToastProvider';
-import PageHeader from '@/components/PageHeader';
-import ConfirmModal from '@/components/ConfirmModal';
-import SSHSetupModal from '@/components/SSHSetupModal';
-import LogLevelControl from '@/components/LogLevelControl';
-import FileViewer from '@/components/FileViewer';
-import { AppConfig } from '@/lib/config';
-import { getNodes, createNode, editNode, deleteNode, setNodeAsDefault } from '@/app/actions/nodes';
-import { checkConnection, checkFullConnection } from '@/app/actions/ssh';
-import { PodmanConnection } from '@/lib/nodes';
-import { useEscapeKey } from '@/hooks/useEscapeKey';
-import type { BackupLogEntry, BackupLogStatus, BackupPreviewResult, BackupRestoreSelection } from '@/lib/systemBackup';
-
-type TemplateSettingsSchemaEntry = {
-  default: string;
-  description?: string;
-  required?: boolean;
-};
-
-const DEFAULT_TEMPLATE_SCHEMA: Record<string, TemplateSettingsSchemaEntry> = {
-  DATA_DIR: {
-    default: '/mnt/data',
-    description: 'Base directory used by all templates for persistent data. Applies to new deployments.',
-    required: true
-  }
-};
-
-interface AppUpdateStatus {
-  hasUpdate: boolean;
-  current: string;
-  latest: {
-    version: string;
-    url: string;
-    date: string;
-    notes: string;
-  } | null;
-  config: {
-    autoUpdate: {
-      enabled: boolean;
-      schedule: string;
-      channel?: 'stable' | 'test' | 'dev';
-    }
-  };
-}
-
-interface SystemBackupEntrySummary {
-  fileName: string;
-  createdAt: string;
-  size: number;
-}
-
-type BackupStreamEvent =
-  | { type: 'log'; entry: BackupLogEntry }
-  | { type: 'done'; backup: SystemBackupEntrySummary }
-  | { type: 'error'; message: string };
-
-type SettingsOverrides = Partial<{
-  templateValues: Record<string, string>;
-  registriesEnabled: boolean;
-  registries: { name: string; url: string; branch?: string }[];
-}> & {
-  email?: Partial<{
-    enabled: boolean;
-    host: string;
-    port: number;
-    secure: boolean;
-    user: string;
-    pass: string;
-    from: string;
-    to: string[];
-  }>;
-};
-
-const LOG_STATUS_BADGES: Record<BackupLogStatus, string> = {
-  info: 'text-slate-600 bg-slate-100 dark:text-slate-300 dark:bg-slate-800',
-  success: 'text-emerald-700 bg-emerald-100 dark:text-emerald-300 dark:bg-emerald-900/30',
-  error: 'text-red-700 bg-red-100 dark:text-red-300 dark:bg-red-900/30',
-  skip: 'text-amber-700 bg-amber-100 dark:text-amber-300 dark:bg-amber-900/30'
-};
-
-const LOG_STATUS_DOTS: Record<BackupLogStatus, string> = {
-  info: 'bg-slate-400',
-  success: 'bg-emerald-500',
-  error: 'bg-red-500',
-  skip: 'bg-amber-500'
-};
-
-const formatBytes = (size: number): string => {
-  if (!Number.isFinite(size) || size <= 0) return '0 B';
-  const units = ['B', 'KB', 'MB', 'GB', 'TB'];
-  let value = size;
-  let unitIndex = 0;
-  while (value >= 1024 && unitIndex < units.length - 1) {
-    value /= 1024;
-    unitIndex++;
-  }
-  const precision = value >= 10 ? 0 : 1;
-  return `${value.toFixed(precision)} ${units[unitIndex]}`;
-};
-
-const QUADLET_EXTENSIONS = ['.container', '.service', '.kube', '.yml', '.yaml', '.pod', '.network', '.volume'];
-
-type ServiceDataFileCategory = 'config' | 'logs' | 'certs' | 'data' | 'other';
-
-const SERVICE_DATA_CATEGORY_LABELS: Record<ServiceDataFileCategory, string> = {
-  config: 'Configuration',
-  logs: 'Logs',
-  certs: 'Certificates',
-  data: 'Database',
-  other: 'Other Files'
-};
-
-const SERVICE_DATA_CATEGORY_ICONS: Record<ServiceDataFileCategory, string> = {
-  config: '⚙️',
-  logs: '📋',
-  certs: '🔒',
-  data: '🗄️',
-  other: '📄'
-};
-
-function categorizeServiceDataFile(filePath: string): ServiceDataFileCategory {
-  const lower = filePath.toLowerCase();
-  const fileName = lower.split('/').pop() || '';
-  const dir = lower.split('/').slice(0, -1).join('/');
-
-  // Config files
-  if (dir.includes('proxy_host') || dir.includes('conf.d') || dir.includes('nginx') && fileName.endsWith('.conf'))
-    return 'config';
-  if (fileName.endsWith('.conf') || fileName.endsWith('.json') || fileName.endsWith('.yml') || fileName.endsWith('.yaml'))
-    return 'config';
-  if (fileName === 'settings' || fileName === 'config' || dir.includes('settings'))
-    return 'config';
-
-  // Logs
-  if (dir.includes('logs') || dir.includes('log') || fileName.endsWith('.log'))
-    return 'logs';
-
-  // Certificates
-  if (dir.includes('letsencrypt') || dir.includes('certs') || dir.includes('ssl') || dir.includes('tls'))
-    return 'certs';
-  if (fileName.endsWith('.pem') || fileName.endsWith('.crt') || fileName.endsWith('.key') || fileName.endsWith('.cert'))
-    return 'certs';
-
-  // Database
-  if (fileName.endsWith('.sqlite') || fileName.endsWith('.db') || fileName.endsWith('.sql'))
-    return 'data';
-  if (dir.includes('database') || dir.includes('db'))
-    return 'data';
-
-  return 'other';
-}
-
-function groupServiceDataFiles(files: string[]): { category: ServiceDataFileCategory; files: string[] }[] {
-  const groups = new Map<ServiceDataFileCategory, string[]>();
-  for (const file of files) {
-    const cat = categorizeServiceDataFile(file);
-    if (!groups.has(cat)) groups.set(cat, []);
-    groups.get(cat)!.push(file);
-  }
-  const order: ServiceDataFileCategory[] = ['config', 'certs', 'data', 'logs', 'other'];
-  return order
-    .filter(cat => groups.has(cat))
-    .map(cat => ({ category: cat, files: groups.get(cat)! }));
-}
-
-function groupFilesByService(files: { relativePath: string; fileName: string }[]): { service: string; files: { relativePath: string; fileName: string }[] }[] {
-  const groups = new Map<string, { relativePath: string; fileName: string }[]>();
-  for (const file of files) {
-    const name = file.fileName;
-    let service = '';
-    // Try to extract a service prefix: strip known quadlet extensions, then take the base name
-    for (const ext of QUADLET_EXTENSIONS) {
-      if (name.endsWith(ext)) {
-        // e.g. "immich-database.container" → "immich-database" → prefix "immich"
-        // e.g. "adguard.kube" → "adguard"
-        const stem = name.slice(0, -ext.length);
-        // Take part before first hyphen as the service group, or the whole stem
-        const dashIdx = stem.indexOf('-');
-        service = dashIdx > 0 ? stem.slice(0, dashIdx) : stem;
-        // Handle dotted names like "korgraph.io.full-stack.service" → "korgraph"
-        const dotIdx = service.indexOf('.');
-        if (dotIdx > 0) service = service.slice(0, dotIdx);
-        break;
-      }
-    }
-    if (!service) service = '_other';
-    if (!groups.has(service)) groups.set(service, []);
-    groups.get(service)!.push(file);
-  }
-  // Sort: named groups first (alphabetical), _other last
-  return Array.from(groups.entries())
-    .sort(([a], [b]) => {
-      if (a === '_other') return 1;
-      if (b === '_other') return -1;
-      return a.localeCompare(b);
-    })
-    .map(([service, files]) => ({ service, files }));
-}
-
-export default function SettingsPage() {
-  const [loading, setLoading] = useState(true);
-  const [saving, setSaving] = useState(false);
-  const [configReady, setConfigReady] = useState(false);
-  const [activeTab, setActiveTab] = useState<'nodes' | 'backups' | 'notifications' | 'updates' | 'advanced'>('nodes');
-  const { addToast } = useToast();
-
-  // App Update State
-  const [appUpdate, setAppUpdate] = useState<AppUpdateStatus | null>(null);
-  const [checkingUpdate, setCheckingUpdate] = useState(false);
-  const [isUpdateModalOpen, setIsUpdateModalOpen] = useState(false);
-  
-  // Update Progress State
-  const [updateStatus, setUpdateStatus] = useState<'idle' | 'updating' | 'restarting' | 'error' | 'success'>('idle');
-  const [updateProgress, setUpdateProgress] = useState(0);
-  const [updateMessage, setUpdateMessage] = useState('');
-  const [updateError, setUpdateError] = useState('');
-
-  // Backup State
-  const [backups, setBackups] = useState<SystemBackupEntrySummary[]>([]);
-  const [backupsLoading, setBackupsLoading] = useState(true);
-  const [creatingBackup, setCreatingBackup] = useState(false);
-  const [restoreOverlayOpen, setRestoreOverlayOpen] = useState(false);
-  const [restoringBackup, setRestoringBackup] = useState(false);
-  const [restorePreview, setRestorePreview] = useState<BackupPreviewResult | null>(null);
-  const [restoreSource, setRestoreSource] = useState<{ type: 'stored' | 'upload'; fileName?: string; token?: string } | null>(null);
-  const [restoreUploadError, setRestoreUploadError] = useState<string | null>(null);
-  const [restoreFilePreview, setRestoreFilePreview] = useState<{ nodeName: string; relativePath: string; content: string; loading: boolean } | null>(null);
-  const [restoreFilePreviewError, setRestoreFilePreviewError] = useState<string | null>(null);
-  const [restoreSelectionState, setRestoreSelectionState] = useState<{
-    nodes: Record<string, boolean>;
-    checks: Record<string, boolean>;
-    configFlags: {
-      externalLinks: boolean;
-      registries: boolean;
-      gateway: boolean;
-      notifications: boolean;
-      templateSettings: boolean;
-      logLevel: boolean;
-      update: boolean;
-    };
-    nodeFiles: Record<string, Record<string, boolean>>;
-    targetNodes: Record<string, string>;
-    serviceData: Record<string, Record<string, boolean>>;
-  } | null>(null);
-  const [backupLog, setBackupLog] = useState<BackupLogEntry[]>([]);
-  const [backupStatus, setBackupStatus] = useState<'idle' | 'running' | 'success' | 'error'>('idle');
-  const [deleteTarget, setDeleteTarget] = useState<SystemBackupEntrySummary | null>(null);
-  const [deletingBackup, setDeletingBackup] = useState(false);
-  const [restoreExpandedSections, setRestoreExpandedSections] = useState<Record<string, boolean>>({});
-
-  // Backup Sync State
-  const [backupSync, setBackupSync] = useState<{
-    enabled: boolean;
-    schedule: 'hourly' | 'daily' | 'weekly' | 'monthly';
-    time: string;
-    dayOfWeek?: number;
-    dayOfMonth?: number;
-    targetType: 'local' | 'ssh' | 'smb' | 'nfs';
-    localPath: string;
-    sshHost: string;
-    sshPort: string;
-    sshUser: string;
-    sshPath: string;
-    sshIdentityFile: string;
-    smbHost: string;
-    smbShare: string;
-    smbPath: string;
-    smbUsername: string;
-    smbPassword: string;
-    nfsHost: string;
-    nfsExport: string;
-    nfsPath: string;
-    sourcePath: string;
-    excludePatterns: string;
-    lastRun?: string;
-    lastStatus?: 'success' | 'error';
-    lastMessage?: string;
-    lastDuration?: number;
-  }>({
-    enabled: false,
-    schedule: 'daily',
-    time: '02:00',
-    targetType: 'local',
-    localPath: '/mnt/backup',
-    sshHost: '', sshPort: '22', sshUser: 'root', sshPath: '/backup', sshIdentityFile: '/app/data/ssh/id_rsa',
-    smbHost: '', smbShare: '', smbPath: '', smbUsername: '', smbPassword: '',
-    nfsHost: '', nfsExport: '', nfsPath: '',
-    sourcePath: '/mnt/data',
-    excludePatterns: '',
-  });
-  const [backupSyncHistory, setBackupSyncHistory] = useState<Array<{ success: boolean; startedAt: string; completedAt: string; duration: number; message: string; filesTransferred?: number }>>([]);
-  const [backupSyncRunning, setBackupSyncRunning] = useState(false);
-  const [backupSyncTesting, setBackupSyncTesting] = useState(false);
-  const [backupSyncSaving, setBackupSyncSaving] = useState(false);
-  const [backupSyncTestResult, setBackupSyncTestResult] = useState<{ success: boolean; message: string } | null>(null);
-
-  // Nginx Config Export/Import State
-  const [nginxExporting, setNginxExporting] = useState(false);
-  const [nginxImporting, setNginxImporting] = useState(false);
-  const [nginxNode, setNginxNode] = useState<string | null>(null);
-  const [nginxInstalled, setNginxInstalled] = useState(false);
-  const nginxFileInputRef = useRef<HTMLInputElement>(null);
-  const [nginxDiag, setNginxDiag] = useState<{ reason: string; debug: string[]; node?: string; confDir?: string } | null>(null);
-  const [nginxDiagExpanded, setNginxDiagExpanded] = useState(false);
-
-  // Email Form State
-  const [emailEnabled, setEmailEnabled] = useState(false);
-  const [emailHost, setEmailHost] = useState('');
-  const [emailPort, setEmailPort] = useState(587);
-  const [emailSecure, setEmailSecure] = useState(false);
-  const [emailUser, setEmailUser] = useState('');
-  const [emailPass, setEmailPass] = useState('');
-  const [emailFrom, setEmailFrom] = useState('');
-  const [emailRecipients, setEmailRecipients] = useState<string[]>([]);
-  const [newRecipient, setNewRecipient] = useState('');
-
-  // Registry State
-  const [registriesEnabled, setRegistriesEnabled] = useState(true);
-  const [registries, setRegistries] = useState<{name: string, url: string, branch?: string}[]>([]);
-  const [newRegName, setNewRegName] = useState('');
-  const [newRegUrl, setNewRegUrl] = useState('');
-  const [newRegBranch, setNewRegBranch] = useState('');
-
-  // Server Name
-  const [serverName, setServerName] = useState('');
-
-  // Template Settings
-  const [templateSchema, setTemplateSchema] = useState<Record<string, TemplateSettingsSchemaEntry>>(DEFAULT_TEMPLATE_SCHEMA);
-  const [templateValues, setTemplateValues] = useState<Record<string, string>>({ DATA_DIR: DEFAULT_TEMPLATE_SCHEMA.DATA_DIR.default });
-  const [newVarKey, setNewVarKey] = useState('');
-  const [newVarValue, setNewVarValue] = useState('');
-
-  const router = useRouter();
-
-  // Nodes State
-  const [nodes, setNodes] = useState<PodmanConnection[]>([]);
-  const [newNodeName, setNewNodeName] = useState('');
-  const [newNodeDest, setNewNodeDest] = useState('');
-  const [newNodeIdentity, setNewNodeIdentity] = useState('/app/data/ssh/id_rsa');
-  const [addingNode, setAddingNode] = useState(false);
-  const [editingNode, setEditingNode] = useState<string | null>(null);
-  const [nodeDraft, setNodeDraft] = useState<{ name: string; destination: string; identity: string }>({
-    name: '',
-    destination: '',
-    identity: '/app/data/ssh/id_rsa'
-  });
-  const [savingNode, setSavingNode] = useState(false);
-  const [nodeHealth, setNodeHealth] = useState<Record<string, { loading: boolean; online: boolean; auth: boolean; error?: string }>>({});
-
-  // SSH Setup Modal
-  const [isSSHModalOpen, setIsSSHModalOpen] = useState(false);
-  const [sshModalDefaults, setSshModalDefaults] = useState({ host: '', port: 22, user: 'root' });
-
-    const normalizeNodeName = useCallback((name: string) => name.trim().toLowerCase(), []);
-    const dedupeNodes = useCallback((list: PodmanConnection[]) => {
-    const seen = new Map<string, PodmanConnection>();
-    for (const node of list) {
-      const key = normalizeNodeName(node.Name);
-      const existing = seen.get(key);
-      if (!existing) {
-        seen.set(key, node);
-      } else {
-        seen.set(key, { ...existing, Default: existing.Default || node.Default });
-      }
-    }
-    const result = Array.from(seen.values());
-    if (result.length > 0 && !result.some(n => n.Default)) {
-      result[0].Default = true;
-    }
-    return result;
-    }, [normalizeNodeName]);
-
-  // Update sshModalDefaults when adding node inputs change, to have valid defaults if user clicks the sidebar button manually
-  useEffect(() => {
-    let host = '', port = 22, user = 'root';
-    try {
-        if (newNodeDest) {
-            const urlStr = newNodeDest.includes('://') ? newNodeDest : `ssh://${newNodeDest}`;
-            const url = new URL(urlStr);
-            host = url.hostname;
-            port = url.port ? parseInt(url.port) : 22;
-            user = url.username || 'root';
-            setSshModalDefaults({ host, port, user });
-        }
-    } catch {
-        // Ignore parse error
-    }
-  }, [newNodeDest]);
-
-  const fetchConfig = useCallback(async () => {
-    try {
-      const [res, updateRes] = await Promise.all([
-        fetch('/api/settings'),
-        fetch('/api/system/update')
-      ]);
-      
-      if (!res.ok) throw new Error('Failed to fetch config');
-      const data: AppConfig = await res.json();
-      
-      if (updateRes.ok) {
-        setAppUpdate(await updateRes.json());
-      } else if (data.autoUpdate) {
-        // Fallback if update API fails but we have config
-        setAppUpdate({
-            hasUpdate: false,
-            current: 'Unknown',
-            latest: null,
-            config: { autoUpdate: data.autoUpdate }
-        });
-      }
-      
-      // Initialize form
-      if (data.registries) {
-        if (Array.isArray(data.registries)) {
-            setRegistries(data.registries);
-            setRegistriesEnabled(true);
-        } else {
-            setRegistries(data.registries.items);
-            setRegistriesEnabled(data.registries.enabled);
-        }
-      }
-
-      setServerName(data.serverName || '');
-
-      const response = data as AppConfig & { templateSettingsSchema?: Record<string, TemplateSettingsSchemaEntry> };
-      const schema = response.templateSettingsSchema || DEFAULT_TEMPLATE_SCHEMA;
-      const defaults = Object.fromEntries(Object.entries(schema).map(([k, v]) => [k, v.default ?? ''])) as Record<string, string>;
-      const persisted = (data as AppConfig).templateSettings || {};
-      const merged = { ...defaults, ...persisted };
-      setTemplateSchema(schema);
-      setTemplateValues(merged);
-
-      // Fetch nodes
-      try {
-        const nodesList = await getNodes();
-        setNodes(dedupeNodes(nodesList));
-      } catch (e) {
-        console.error('Failed to fetch nodes', e);
-      }
-
-      if (data.notifications?.email) {
-        const e = data.notifications.email;
-        setEmailEnabled(e.enabled);
-        setEmailHost(e.host);
-        setEmailPort(e.port);
-        setEmailSecure(e.secure);
-        setEmailUser(e.user);
-        setEmailPass(e.pass);
-        setEmailFrom(e.from);
-        setEmailRecipients(e.to || []);
-      }
-
-      setConfigReady(true);
-    } catch (error) {
-      console.error(error);
-      addToast('error', 'Failed to load settings');
-    } finally {
-      setLoading(false);
-    }
-  }, [addToast, dedupeNodes]);
-
-  useEffect(() => {
-    fetchConfig();
-  }, [fetchConfig]);
-
-  const fetchBackups = useCallback(async () => {
-    setBackupsLoading(true);
-    try {
-      const res = await fetch('/api/settings/backups');
-      if (!res.ok) {
-        const errorData = await res.json().catch(() => ({}));
-        throw new Error(errorData.error || 'Unable to load backups');
-      }
-      const data: SystemBackupEntrySummary[] = await res.json();
-      setBackups(data);
-    } catch (error) {
-      console.error(error);
-      const message = error instanceof Error ? error.message : undefined;
-      addToast('error', 'Failed to load backups', message);
-    } finally {
-      setBackupsLoading(false);
-    }
-  }, [addToast]);
-
-  useEffect(() => {
-    fetchBackups();
-  }, [fetchBackups]);
-
-  // Fetch Backup Sync config
-  const fetchBackupSync = useCallback(async () => {
-    try {
-      const res = await fetch('/api/settings/backup-sync');
-      if (!res.ok) return;
-      const data = await res.json();
-      if (data.config) {
-        const c = data.config;
-        const t = c.target || { type: 'local', path: '/mnt/backup' };
-        setBackupSync(prev => ({
-          ...prev,
-          enabled: c.enabled ?? false,
-          schedule: c.schedule ?? 'daily',
-          time: c.time ?? '02:00',
-          dayOfWeek: c.dayOfWeek,
-          dayOfMonth: c.dayOfMonth,
-          sourcePath: c.sourcePath ?? '/mnt/data',
-          excludePatterns: (c.excludePatterns || []).join('\n'),
-          targetType: t.type ?? 'local',
-          localPath: t.type === 'local' ? t.path : prev.localPath,
-          sshHost: t.type === 'ssh' ? t.host : prev.sshHost,
-          sshPort: t.type === 'ssh' ? String(t.port ?? 22) : prev.sshPort,
-          sshUser: t.type === 'ssh' ? t.user : prev.sshUser,
-          sshPath: t.type === 'ssh' ? t.path : prev.sshPath,
-          sshIdentityFile: t.type === 'ssh' ? (t.identityFile ?? '/app/data/ssh/id_rsa') : prev.sshIdentityFile,
-          smbHost: t.type === 'smb' ? t.host : prev.smbHost,
-          smbShare: t.type === 'smb' ? t.share : prev.smbShare,
-          smbPath: t.type === 'smb' ? (t.path ?? '') : prev.smbPath,
-          smbUsername: t.type === 'smb' ? (t.username ?? '') : prev.smbUsername,
-          smbPassword: t.type === 'smb' ? (t.password ?? '') : prev.smbPassword,
-          nfsHost: t.type === 'nfs' ? t.host : prev.nfsHost,
-          nfsExport: t.type === 'nfs' ? t.export : prev.nfsExport,
-          nfsPath: t.type === 'nfs' ? (t.path ?? '') : prev.nfsPath,
-          lastRun: c.lastRun,
-          lastStatus: c.lastStatus,
-          lastMessage: c.lastMessage,
-          lastDuration: c.lastDuration,
-        }));
-      }
-      if (data.history) setBackupSyncHistory(data.history);
-      setBackupSyncRunning(data.running ?? false);
-    } catch { /* ignore */ }
-  }, []);
-
-  useEffect(() => {
-    fetchBackupSync();
-  }, [fetchBackupSync]);
-
-  const persistSettings = useCallback(async (overrides?: SettingsOverrides) => {
-    if (!configReady || saving) return;
-    setSaving(true);
-    try {
-      const templateDefaults = Object.fromEntries(
-        Object.entries(templateSchema).map(([key, meta]) => [key, meta.default ?? ''])
-      ) as Record<string, string>;
-      const effectiveTemplateValues = overrides?.templateValues ?? templateValues;
-      const enforcedTemplateValues = {
-        ...templateDefaults,
-        ...effectiveTemplateValues
-      } as Record<string, string>;
-
-      const effectiveRegistries = overrides?.registries ?? registries;
-      const effectiveRegistriesEnabled = overrides?.registriesEnabled ?? registriesEnabled;
-
-      const emailOverrides = overrides?.email ?? {};
-      const emailConfig = {
-        enabled: emailOverrides.enabled ?? emailEnabled,
-        host: emailOverrides.host ?? emailHost,
-        port: emailOverrides.port ?? emailPort,
-        secure: emailOverrides.secure ?? emailSecure,
-        user: emailOverrides.user ?? emailUser,
-        pass: emailOverrides.pass ?? emailPass,
-        from: emailOverrides.from ?? emailFrom,
-        to: emailOverrides.to ?? emailRecipients
-      };
-
-      const newConfig: Partial<AppConfig> = {
-        serverName: serverName || undefined,
-        templateSettings: enforcedTemplateValues,
-        registries: {
-          enabled: effectiveRegistriesEnabled,
-          items: effectiveRegistries
-        },
-        notifications: {
-          email: emailConfig
-        }
-      };
-
-      const res = await fetch('/api/settings', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(newConfig)
-      });
-
-      if (!res.ok) {
-        const errorBody = await res.json().catch(() => ({}));
-        throw new Error(errorBody.error || 'Failed to save settings');
-      }
-
-      addToast('success', 'Settings saved', 'Your changes were stored.');
-    } catch (error) {
-      console.error(error);
-      addToast('error', 'Failed to save settings', error instanceof Error ? error.message : undefined);
-    } finally {
-      setSaving(false);
-    }
-  }, [addToast, configReady, emailEnabled, emailFrom, emailHost, emailPass, emailPort, emailRecipients, emailSecure, emailUser, registries, registriesEnabled, saving, serverName, templateSchema, templateValues]);
-
-  const handleAddRecipient = () => {
-    if (newRecipient && !emailRecipients.includes(newRecipient)) {
-      const updatedRecipients = [...emailRecipients, newRecipient];
-      setEmailRecipients(updatedRecipients);
-      setNewRecipient('');
-      void persistSettings({ email: { to: updatedRecipients } });
-    }
-  };
-
-  const handleRemoveRecipient = (email: string) => {
-    const updatedRecipients = emailRecipients.filter(e => e !== email);
-    setEmailRecipients(updatedRecipients);
-    void persistSettings({ email: { to: updatedRecipients } });
-  };
-
-  const handleEmailEnabledToggle = (enabled: boolean) => {
-    setEmailEnabled(enabled);
-    void persistSettings({ email: { enabled } });
-  };
-
-  const handleEmailSecureToggle = (secure: boolean) => {
-    setEmailSecure(secure);
-    void persistSettings({ email: { secure } });
-  };
-
-  const handleAddNode = () => {
-    void submitNode('create', {
-      name: newNodeName.trim(),
-      destination: newNodeDest.trim(),
-      identity: newNodeIdentity.trim()
-    });
-  };
-
-  const startEditingNode = (node: PodmanConnection) => {
-    setEditingNode(node.Name);
-    setNodeDraft({ name: node.Name, destination: node.URI, identity: node.Identity });
-  };
-
-  const cancelInlineEdit = () => {
-    setEditingNode(null);
-    setNodeDraft({ name: '', destination: '', identity: '/app/data/ssh/id_rsa' });
-  };
-
-  const handleInlineSave = () => {
-    if (!editingNode) return;
-    void submitNode('edit', {
-      originalName: editingNode,
-      name: nodeDraft.name.trim(),
-      destination: nodeDraft.destination.trim(),
-      identity: nodeDraft.identity.trim()
-    });
-  };
-
-  const parseDestination = (destination: string) => {
-    let host = '', port = 22, user = 'root';
-    try {
-      const urlStr = destination.includes('://') ? destination : `ssh://${destination}`;
-      const parsed = new URL(urlStr);
-      host = parsed.hostname;
-      port = parsed.port ? parseInt(parsed.port) : 22;
-      user = parsed.username || 'root';
-    } catch {
-      // ignore
-    }
-    return { host, port, user };
-  };
-
-  const handleDeploySSHKey = (node: PodmanConnection) => {
-    const { host, port, user } = parseDestination(node.URI);
-    setSshModalDefaults({ host, port, user });
-    setIsSSHModalOpen(true);
-  };
-
-  const handleOpenTerminal = (node: PodmanConnection) => {
-    const target = encodeURIComponent(node.Name);
-    router.push(`/terminal?node=${target}`);
-  };
-
-  const submitNode = useCallback(async (mode: 'create' | 'edit', payload: { name: string; destination: string; identity: string; originalName?: string }) => {
-    if (!payload.name || !payload.destination || !payload.identity) return false;
-
-    const { host, port, user } = parseDestination(payload.destination);
-    const setBusy = mode === 'create' ? setAddingNode : setSavingNode;
-    setBusy(true);
-
-    const normalizedName = normalizeNodeName(payload.name);
-    if (normalizedName === 'local') {
-      addToast('error', 'Local is reserved', 'The Local node is managed automatically. Please use a different name.');
-      setBusy(false);
-      return false;
-    }
-
-    if (host) {
-      const check = await checkConnection(host, port);
-      if (!check.success || !check.isOpen) {
-        addToast('error', 'Connection Failed', `Could not connect to ${host}:${port}. Is the server reachable?`);
-        setBusy(false);
-        return false;
-      }
-    }
-
-    try {
-      const result = mode === 'edit' && payload.originalName
-        ? await editNode(payload.originalName, payload.name, payload.destination, payload.identity)
-        : await createNode(payload.name, payload.destination, payload.identity);
-
-      if (result.success) {
-        setNodes(dedupeNodes(await getNodes()));
-        if (mode === 'create') {
-          setNewNodeName('');
-          setNewNodeDest('');
-          setNewNodeIdentity('/app/data/ssh/id_rsa');
-        } else {
-          setEditingNode(null);
-          setNodeDraft({ name: '', destination: '', identity: '/app/data/ssh/id_rsa' });
-        }
-
-        const warning = (result as { warning?: string }).warning;
-        if (warning) {
-          if (warning.includes('timed out') || warning.includes('Permission denied') || warning.includes('password') || warning.includes('publickey')) {
-            addToast('warning', 'SSH Connection Failed', 'The node was saved, but we could not connect. It seems password-less SSH is not configured.');
-            if (host) {
-              setSshModalDefaults({ host, port, user });
-              setIsSSHModalOpen(true);
-            }
-          } else {
-            addToast('warning', mode === 'edit' ? 'Node updated with warning' : 'Node added with warning', warning);
-          }
-        } else {
-          addToast('success', mode === 'edit' ? 'Node updated' : 'Node added');
-        }
-
-        return true;
-      }
-
-      addToast('error', mode === 'edit' ? 'Failed to update node' : 'Failed to add node', (result as { error?: string }).error);
-      return false;
-    } catch (error) {
-      addToast('error', mode === 'edit' ? 'Failed to update node' : 'Failed to add node', error instanceof Error ? error.message : String(error));
-      return false;
-    } finally {
-      setBusy(false);
-    }
-  }, [addToast, dedupeNodes, normalizeNodeName, setSshModalDefaults]);
-
-  const handleAddRegistry = () => {
-    if (newRegName && newRegUrl) {
-      const updatedRegistries = [...registries, { name: newRegName, url: newRegUrl, branch: newRegBranch || undefined }];
-      setRegistries(updatedRegistries);
-      setNewRegName('');
-      setNewRegUrl('');
-      setNewRegBranch('');
-      void persistSettings({ registries: updatedRegistries });
-    }
-  };
-
-  const handleRemoveRegistry = (name: string) => {
-    const updatedRegistries = registries.filter(r => r.name !== name);
-    setRegistries(updatedRegistries);
-    void persistSettings({ registries: updatedRegistries });
-  };
-
-  const handleRegistriesToggle = (enabled: boolean) => {
-    setRegistriesEnabled(enabled);
-    void persistSettings({ registriesEnabled: enabled });
-  };
-
-  const handleTemplateValueChange = (key: string, value: string) => {
-    setTemplateValues(prev => ({ ...prev, [key]: value }));
-  };
-
-  const handleAddTemplateVariable = () => {
-    if (!newVarKey.trim()) return;
-    const key = newVarKey.trim();
-    const updated = { ...templateValues, [key]: newVarValue };
-    setTemplateValues(updated);
-    setNewVarKey('');
-    setNewVarValue('');
-    void persistSettings({ templateValues: updated });
-  };
-
-  const handleRemoveTemplateVariable = (key: string) => {
-    const meta = templateSchema[key];
-    if (meta?.required) return;
-    setTemplateValues(prev => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const { [key]: _removed, ...rest } = prev;
-      void persistSettings({ templateValues: rest });
-      return rest;
-    });
-  };
-
-  const checkHealth = useCallback(async (nodeName: string) => {
-    setNodeHealth(prev => ({ ...prev, [nodeName]: { loading: true, online: false, auth: false } }));
-    
-    const node = nodes.find(n => n.Name === nodeName);
-    if (!node) return;
-    
-    if (node.URI === 'local') {
-         setNodeHealth(prev => ({
-           ...prev,
-           [nodeName]: {
-             loading: false,
-             online: false,
-             auth: false,
-             error: 'Legacy local nodes are unsupported. Edit this node to use ssh://user@host.'
-           }
-         }));
-         return;
-    }
-
-    try {
-      let host = '', port = 22, user = 'root';
-      
-      if (node.URI.startsWith('ssh://')) {
-          const url = new URL(node.URI);
-          host = url.hostname;
-          port = url.port ? parseInt(url.port) : 22;
-          user = url.username || 'root';
-      } else {
-           const parts = node.URI.split('@');
-           if (parts.length === 2) {
-               user = parts[0];
-               host = parts[1];
-           } else {
-               host = node.URI;
-           }
-      }
-
-      const res = await checkFullConnection(host, port, user, node.Identity);
-      
-      setNodeHealth(prev => ({ 
-         ...prev, 
-         [nodeName]: { 
-             loading: false, 
-             online: res.success || (res.stage === 'auth'), 
-             auth: res.success,
-             error: res.error 
-         } 
-      }));
-    } catch (e) {
-       setNodeHealth(prev => ({ ...prev, [nodeName]: { loading: false, online: false, auth: false, error: String(e) } }));
-    }
-  }, [nodes]);
-
-  useEffect(() => {
-    if (nodes.length > 0) {
-        nodes.forEach(n => {
-            // Only check if not already checked or if previously failed/loading (optional)
-            // Ideally only check on mount or when nodes change
-            if (!nodeHealth[n.Name]) checkHealth(n.Name);
-        });
-    }
-  }, [nodes, checkHealth, nodeHealth]);
-
-  const handleDeleteNode = async (name: string) => {
-    const res = await deleteNode(name);
-    if (res.success) {
-      setNodes(dedupeNodes(await getNodes()));
-      addToast('success', 'Node removed');
-    } else {
-      addToast('error', 'Failed to remove node');
-    }
-  };
-
-  const handleSetDefaultNode = async (name: string) => {
-    const res = await setNodeAsDefault(name);
-    if (res.success) {
-      setNodes(dedupeNodes(await getNodes()));
-      addToast('success', 'Default node updated');
-    } else {
-      addToast('error', 'Failed to set default node');
-    }
-  };
-
-  const handleCheckUpdate = async () => {
-    setCheckingUpdate(true);
-    try {
-        const res = await fetch('/api/system/update');
-        if (res.ok) {
-            const data = await res.json();
-            setAppUpdate(data);
-            const latestVer = data.latest?.version || 'Unknown';
-            if (data.hasUpdate) {
-                addToast('success', 'Update available', `Version ${latestVer} is available.`);
-            } else {
-                addToast('info', 'System up to date', `You are using the latest version (${latestVer}).`);
-            }
-        } else {
-             const errorData = await res.json().catch(() => ({}));
-             const errorMessage = errorData.error || `Server error (${res.status})`;
-             throw new Error(errorMessage);
-        }
-    } catch (e) {
-        console.error(e);
-        const msg = e instanceof Error ? e.message : 'Unknown error';
-        addToast('error', 'Update Check Failed', msg);
-    } finally {
-        setCheckingUpdate(false);
-    }
-  };
-
-  const handleAppUpdate = () => {
-    if (!appUpdate?.latest) return;
-    setIsUpdateModalOpen(true);
-  };
-
-  const confirmAppUpdate = async () => {
-    if (!appUpdate?.latest) return;
-    
-    setIsUpdateModalOpen(false);
-    setUpdateStatus('updating');
-    setUpdateProgress(0);
-    setUpdateMessage('Initializing update...');
-    
-    try {
-      // Connect to socket for progress updates
-      const { io } = await import('socket.io-client');
-      const socket = io();
-
-      socket.on('update:progress', (data: { step: string, progress: number, message: string }) => {
-          setUpdateProgress(data.progress);
-          setUpdateMessage(data.message);
-          if (data.step === 'restart') {
-              setUpdateStatus('restarting');
-              setUpdateMessage('Service is restarting. This page will reload automatically...');
-              
-              // Wait a bit then try to reload
-              setTimeout(() => {
-                  window.location.reload();
-              }, 5000);
-          }
-      });
-
-      socket.on('update:error', (data: { error: string }) => {
-          setUpdateStatus('error');
-          setUpdateError(data.error);
-          socket.disconnect();
-      });
-
-      const res = await fetch('/api/system/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ action: 'update', version: appUpdate.latest.version })
-      });
-      
-      if (!res.ok) {
-        throw new Error('Update failed to start');
-      }
-    } catch (e) {
-      setUpdateStatus('error');
-      setUpdateError(e instanceof Error ? e.message : 'Unknown error');
-    }
-  };
-
-  const toggleAutoUpdate = async () => {
-    if (!appUpdate) return;
-    const newState = !appUpdate.config.autoUpdate.enabled;
-    
-    try {
-      const res = await fetch('/api/system/update', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ 
-            action: 'configure', 
-            autoUpdate: { enabled: newState } 
-        })
-      });
-      
-      if (res.ok) {
-        const data = await res.json();
-        setAppUpdate(prev => prev ? { ...prev, config: data.config } : null);
-        addToast('success', 'Settings saved', `Auto-update ${newState ? 'enabled' : 'disabled'}.`);
-      }
-    } catch {
-      addToast('error', 'Error', 'Failed to save settings.');
-    }
-  };
-
-  // ─── Backup Sync Handlers ─────────────────────────────────────────
-  const buildBackupTarget = () => {
-    const s = backupSync;
-    switch (s.targetType) {
-      case 'local': return { type: 'local' as const, path: s.localPath };
-      case 'ssh': return { type: 'ssh' as const, host: s.sshHost, port: parseInt(s.sshPort) || 22, user: s.sshUser, path: s.sshPath, identityFile: s.sshIdentityFile || undefined };
-      case 'smb': return { type: 'smb' as const, host: s.smbHost, share: s.smbShare, path: s.smbPath || undefined, username: s.smbUsername || undefined, password: s.smbPassword || undefined };
-      case 'nfs': return { type: 'nfs' as const, host: s.nfsHost, export: s.nfsExport, path: s.nfsPath || undefined };
-    }
-  };
-
-  const handleSaveBackupSync = async () => {
-    setBackupSyncSaving(true);
-    try {
-      const config = {
-        enabled: backupSync.enabled,
-        schedule: backupSync.schedule,
-        time: backupSync.time,
-        dayOfWeek: backupSync.schedule === 'weekly' ? backupSync.dayOfWeek : undefined,
-        dayOfMonth: backupSync.schedule === 'monthly' ? backupSync.dayOfMonth : undefined,
-        target: buildBackupTarget(),
-        sourcePath: backupSync.sourcePath,
-        excludePatterns: backupSync.excludePatterns.split('\n').map(s => s.trim()).filter(Boolean),
-      };
-      const res = await fetch('/api/settings/backup-sync', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ action: 'save', config }),
-      });
-      if (!res.ok) throw new Error((await res.json()).error || 'Save failed');
-      addToast('success', 'Backup Sync', 'Configuration saved.');
-      await fetchBackupSync();
-    } catch (e) {
-      addToast('error', 'Save Failed', e instanceof Error ? e.message : 'Unknown error');
-    } finally {
-      setBackupSyncSaving(false);
-    }
-  };
-
-  const handleTestBackupSync = async () => {
-    setBackupSyncTesting(true);
-    setBackupSyncTestResult(null);
-    try {
-      const res = await fetch('/api/settings/backup-sync', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ action: 'test', target: buildBackupTarget() }),
-      });
-      const result = await res.json();
-      setBackupSyncTestResult(result);
-    } catch (e) {
-      setBackupSyncTestResult({ success: false, message: e instanceof Error ? e.message : 'Connection test failed' });
-    } finally {
-      setBackupSyncTesting(false);
-    }
-  };
-
-  const handleRunBackupSync = async () => {
-    setBackupSyncRunning(true);
-    try {
-      const res = await fetch('/api/settings/backup-sync', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ action: 'run' }),
-      });
-      if (!res.ok) throw new Error((await res.json()).error || 'Run failed');
-      addToast('info', 'Backup', 'Backup sync started. This may take a while.');
-      // Poll for completion
-      const poll = setInterval(async () => {
-        try {
-          const r = await fetch('/api/settings/backup-sync');
-          const data = await r.json();
-          if (!data.running) {
-            clearInterval(poll);
-            setBackupSyncRunning(false);
-            setBackupSyncHistory(data.history || []);
-            if (data.config?.lastStatus === 'success') {
-              addToast('success', 'Backup', data.config.lastMessage || 'Backup completed');
-            } else if (data.config?.lastStatus === 'error') {
-              addToast('error', 'Backup', data.config.lastMessage || 'Backup failed');
-            }
-            setBackupSync(prev => ({
-              ...prev,
-              lastRun: data.config?.lastRun,
-              lastStatus: data.config?.lastStatus,
-              lastMessage: data.config?.lastMessage,
-              lastDuration: data.config?.lastDuration,
-            }));
-          }
-        } catch { /* ignore */ }
-      }, 5000);
-    } catch (e) {
-      setBackupSyncRunning(false);
-      addToast('error', 'Backup Failed', e instanceof Error ? e.message : 'Unknown error');
-    }
-  };
-
-  const handleCreateBackup = async () => {
-    if (creatingBackup) return;
-    setCreatingBackup(true);
-    setBackupStatus('running');
-    setBackupLog([]);
-    let sawDone = false;
-    let errorMessage: string | null = null;
-
-    try {
-      const res = await fetch('/api/settings/backups', { method: 'POST' });
-
-      if (!res.body) {
-        const errorData = await res.json().catch(() => ({}));
-        throw new Error(errorData.error || 'Streaming not supported by server');
-      }
-
-      const reader = res.body.getReader();
-      const decoder = new TextDecoder();
-      let buffer = '';
-
-      while (true) {
-        const { value, done } = await reader.read();
-        buffer += decoder.decode(value ?? new Uint8Array(), { stream: !done });
-
-        let newlineIndex = buffer.indexOf('\n');
-        while (newlineIndex !== -1) {
-          const line = buffer.slice(0, newlineIndex).trim();
-          buffer = buffer.slice(newlineIndex + 1);
-          if (line.length > 0) {
-            try {
-              const event = JSON.parse(line) as BackupStreamEvent;
-              if (event.type === 'log' && event.entry) {
-                setBackupLog(prev => [...prev, event.entry]);
-              } else if (event.type === 'done') {
-                if (!sawDone) {
-                  sawDone = true;
-                  setBackupStatus('success');
-                  addToast('success', 'Backup created', `Archive ${event.backup.fileName} is ready.`);
-                  await fetchBackups();
-                }
-              } else if (event.type === 'error') {
-                errorMessage = event.message || 'Backup failed';
-                setBackupStatus('error');
-                addToast('error', 'Failed to create backup', errorMessage);
-              }
-            } catch {
-              // ignore malformed chunk
-            }
-          }
-          newlineIndex = buffer.indexOf('\n');
-        }
-
-        if (done) break;
-      }
-
-      if (!sawDone && !errorMessage) {
-        throw new Error('Backup stream ended unexpectedly');
-      }
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Unknown error';
-      errorMessage = message;
-      setBackupStatus('error');
-      addToast('error', 'Failed to create backup', message);
-    } finally {
-      setCreatingBackup(false);
-    }
-  };
-
-  const handleDownloadBackup = (fileName: string) => {
-    const link = document.createElement('a');
-    link.href = `/api/settings/backups/download?file=${encodeURIComponent(fileName)}`;
-    link.target = '_blank';
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  };
-
-  // Nginx config export/import
-  const nginxNodeQuery = nginxNode && nginxNode !== 'Local' ? `?node=${encodeURIComponent(nginxNode)}` : '';
-
-  const checkNginxStatus = useCallback(async () => {
-    try {
-      const res = await fetch('/api/system/nginx/status');
-      const data = await res.json();
-      setNginxInstalled(data.installed ?? false);
-      if (data.node) setNginxNode(data.node);
-    } catch {
-      setNginxInstalled(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    checkNginxStatus();
-  }, [checkNginxStatus]);
-
-  const handleNginxExport = async () => {
-    setNginxExporting(true);
-    setNginxDiag(null);
-    try {
-      const res = await fetch(`/api/system/nginx/export${nginxNodeQuery}`);
-      if (!res.ok) throw new Error('Export failed');
-      const data = await res.json();
-      if (!data.files || Object.keys(data.files).length === 0) {
-        setNginxDiag({
-          reason: data.reason || 'No config files found (unknown reason)',
-          debug: data.debug || [],
-          node: data.node,
-          confDir: data.confDir
-        });
-        setNginxDiagExpanded(false);
-        return;
-      }
-      const blob = new Blob([JSON.stringify(data.files, null, 2)], { type: 'application/json' });
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = `nginx-config-${new Date().toISOString().slice(0, 10)}.json`;
-      a.click();
-      URL.revokeObjectURL(url);
-      addToast('success', `Exported ${Object.keys(data.files).length} config file(s)`);
-    } catch {
-      addToast('error', 'Failed to export nginx config');
-    } finally {
-      setNginxExporting(false);
-    }
-  };
-
-  const handleNginxImport = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setNginxImporting(true);
-    setNginxDiag(null);
-    try {
-      const isBackup = file.name.endsWith('.tar.gz') || file.name.endsWith('.tgz');
-      let res: Response;
-
-      if (isBackup) {
-        const formData = new FormData();
-        formData.append('file', file);
-        res = await fetch(`/api/system/nginx/import${nginxNodeQuery}`, {
-          method: 'POST',
-          body: formData
-        });
-      } else {
-        const text = await file.text();
-        const files = JSON.parse(text);
-        if (typeof files !== 'object' || Array.isArray(files)) {
-          throw new Error('Invalid format');
-        }
-        res = await fetch(`/api/system/nginx/import${nginxNodeQuery}`, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ files })
-        });
-      }
-
-      const data = await res.json();
-      if (!res.ok) {
-        setNginxDiag({
-          reason: data.error || 'Import failed (unknown error)',
-          debug: data.debug || [],
-          node: data.node
-        });
-        setNginxDiagExpanded(false);
-        return;
-      }
-      addToast('success', `Imported ${data.imported?.length || 0} config file(s)`);
-    } catch {
-      setNginxDiag({
-        reason: 'Failed to parse the uploaded file. Expected a JSON export file ({ "name.conf": "content" }) or a full ServiceBay backup (.tar.gz).',
-        debug: []
-      });
-      setNginxDiagExpanded(false);
-    } finally {
-      setNginxImporting(false);
-      if (nginxFileInputRef.current) nginxFileInputRef.current.value = '';
-    }
-  };
-
-  const buildDefaultRestoreState = useCallback((preview: BackupPreviewResult) => {
-    const nodesState = Object.fromEntries(preview.config.nodes.map(node => [node.name, true]));
-    const checksState = Object.fromEntries(preview.config.checks.map(check => [check.id, true]));
-    const nodeFilesState: Record<string, Record<string, boolean>> = {};
-    const targetNodes: Record<string, string> = {};
-
-    const availableTargets = ['Local', ...nodes.map(node => node.Name)];
-
-    preview.nodeFiles.forEach(group => {
-      nodeFilesState[group.nodeName] = Object.fromEntries(group.files.map(file => [file.relativePath, true]));
-      targetNodes[group.nodeName] = availableTargets.includes(group.nodeName) ? group.nodeName : 'Local';
-    });
-
-    const serviceDataState: Record<string, Record<string, boolean>> = {};
-    (preview.serviceData || []).forEach(sd => {
-      serviceDataState[sd.name] = Object.fromEntries(sd.files.map(f => [f, true]));
-    });
-
-    setRestoreSelectionState({
-      nodes: nodesState,
-      checks: checksState,
-      configFlags: {
-        externalLinks: preview.config.externalLinks.length > 0,
-        registries: preview.config.registries.length > 0,
-        gateway: Boolean(preview.config.gateway),
-        notifications: Boolean(preview.config.notifications),
-        templateSettings: preview.config.templateSettings.length > 0,
-        logLevel: Boolean(preview.config.logLevel),
-        update: Boolean(preview.config.update)
-      },
-      nodeFiles: nodeFilesState,
-      targetNodes,
-      serviceData: serviceDataState
-    });
-  }, [nodes]);
-
-  const openRestoreOverlay = (reset: boolean = false) => {
-    if (reset) {
-      setRestorePreview(null);
-      setRestoreSource(null);
-      setRestoreSelectionState(null);
-      setRestoreUploadError(null);
-      setRestoreFilePreview(null);
-      setRestoreFilePreviewError(null);
-    }
-    setRestoreOverlayOpen(true);
-    setRestoreUploadError(null);
-    setRestoreExpandedSections({});
-  };
-
-  const closeRestoreOverlay = useCallback(() => {
-    if (restoringBackup) return;
-    setRestoreOverlayOpen(false);
-    setRestorePreview(null);
-    setRestoreSource(null);
-    setRestoreUploadError(null);
-    setRestoreSelectionState(null);
-    setRestoreFilePreview(null);
-    setRestoreFilePreviewError(null);
-  }, [restoringBackup]);
-
-  useEscapeKey(closeRestoreOverlay, restoreOverlayOpen, true);
-  useEscapeKey(() => setRestoreFilePreview(null), Boolean(restoreFilePreview), true);
-
-  const handleRestorePreviewRequest = async (payload: { file?: File; fileName?: string }) => {
-    setRestoreUploadError(null);
-    setRestorePreview(null);
-    setRestoreSource(null);
-    setRestoreSelectionState(null);
-    setRestoreFilePreview(null);
-    setRestoreFilePreviewError(null);
-
-    try {
-      let response: Response;
-      if (payload.file) {
-        const formData = new FormData();
-        formData.append('file', payload.file);
-        response = await fetch('/api/settings/backups/preview', { method: 'POST', body: formData });
-      } else if (payload.fileName) {
-        response = await fetch('/api/settings/backups/preview', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ fileName: payload.fileName })
-        });
-      } else {
-        throw new Error('No backup selected');
-      }
-
-      const data = await response.json().catch(() => ({}));
-      if (!response.ok) {
-        throw new Error(data.error || 'Unable to read backup');
-      }
-
-      setRestorePreview(data.preview as BackupPreviewResult);
-      setRestoreSource(data.source);
-      buildDefaultRestoreState(data.preview as BackupPreviewResult);
-      openRestoreOverlay(false);
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Failed to load backup preview';
-      setRestoreUploadError(message);
-      addToast('error', 'Restore preview failed', message);
-    }
-  };
-
-  const handleRestoreFilePreview = useCallback(async (nodeName: string, relativePath: string) => {
-    if (!restoreSource) return;
-    setRestoreFilePreview({ nodeName, relativePath, content: '', loading: true });
-    setRestoreFilePreviewError(null);
-    try {
-      const payload = restoreSource.type === 'stored'
-        ? { fileName: restoreSource.fileName, nodeName, relativePath }
-        : { uploadToken: restoreSource.token, nodeName, relativePath };
-      const res = await fetch('/api/settings/backups/file', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(payload)
-      });
-      const data = await res.json().catch(() => ({}));
-      if (!res.ok) {
-        throw new Error(data.error || 'Unable to load file');
-      }
-      setRestoreFilePreview({ nodeName, relativePath, content: data.content ?? '', loading: false });
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Unable to load file preview';
-      setRestoreFilePreviewError(message);
-      setRestoreFilePreview({ nodeName, relativePath, content: '', loading: false });
-    }
-  }, [restoreSource]);
-
-  const resolveFilePreviewLanguage = (fileName: string) => {
-    if (fileName.endsWith('.yml') || fileName.endsWith('.yaml')) return 'yaml';
-    if (fileName.endsWith('.kube') || fileName.endsWith('.container') || fileName.endsWith('.pod') || fileName.endsWith('.network') || fileName.endsWith('.volume')) return 'ini';
-    if (fileName.endsWith('.json')) return 'json';
-    if (fileName.endsWith('.sh')) return 'bash';
-    return 'bash';
-  };
-
-  const handleRestoreRequest = (entry: SystemBackupEntrySummary) => {
-    void handleRestorePreviewRequest({ fileName: entry.fileName });
-  };
-
-  const handleRestoreFromFile = (file: File | null) => {
-    if (!file) return;
-    void handleRestorePreviewRequest({ file });
-  };
-
-  const confirmRestoreBackup = useCallback(async () => {
-    if (!restorePreview || !restoreSource || !restoreSelectionState || restoringBackup) return;
-    setRestoringBackup(true);
-    try {
-      const selectedNodes = Object.entries(restoreSelectionState.nodes).filter(([, enabled]) => enabled).map(([name]) => name);
-      const selectedChecks = Object.entries(restoreSelectionState.checks).filter(([, enabled]) => enabled).map(([id]) => id);
-      const nodeFiles = Object.entries(restoreSelectionState.nodeFiles)
-        .map(([sourceNode, filesMap]) => {
-          const files = Object.entries(filesMap).filter(([, enabled]) => enabled).map(([path]) => path);
-          const targetNode = restoreSelectionState.targetNodes[sourceNode];
-          return { sourceNode, targetNode, files };
-        })
-        .filter(group => group.files.length > 0 && group.targetNode);
-
-      const selectedServiceData: { name: string; files?: string[] }[] = [];
-      for (const [name, filesMap] of Object.entries(restoreSelectionState.serviceData)) {
-        const selectedFiles = Object.entries(filesMap).filter(([, enabled]) => enabled).map(([f]) => f);
-        if (selectedFiles.length === 0) continue;
-        const sdPreview = restorePreview.serviceData?.find(sd => sd.name === name);
-        // If all files selected, omit files array (restore entire directory)
-        if (sdPreview && selectedFiles.length === sdPreview.files.length) {
-          selectedServiceData.push({ name });
-        } else {
-          selectedServiceData.push({ name, files: selectedFiles });
-        }
-      }
-
-      const selection: BackupRestoreSelection = {
-        config: {
-          nodes: selectedNodes,
-          checks: selectedChecks,
-          externalLinks: restoreSelectionState.configFlags.externalLinks,
-          registries: restoreSelectionState.configFlags.registries,
-          gateway: restoreSelectionState.configFlags.gateway,
-          notifications: restoreSelectionState.configFlags.notifications,
-          templateSettings: restoreSelectionState.configFlags.templateSettings,
-          logLevel: restoreSelectionState.configFlags.logLevel,
-          update: restoreSelectionState.configFlags.update
-        },
-        nodeFiles,
-        serviceData: selectedServiceData.length > 0 ? selectedServiceData : undefined
-      };
-
-      const payload = restoreSource.type === 'stored'
-        ? { fileName: restoreSource.fileName, selection }
-        : { uploadToken: restoreSource.token, selection };
-
-      const res = await fetch('/api/settings/backups/restore', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(payload)
-      });
-
-      if (!res.ok) {
-        const errorData = await res.json().catch(() => ({}));
-        throw new Error(errorData.error || 'Unable to restore backup');
-      }
-
-      addToast('success', 'Restore complete', 'Selected settings and files were restored.');
-      await fetchBackups();
-      closeRestoreOverlay();
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Restore failed';
-      addToast('error', 'Restore failed', message);
-    } finally {
-      setRestoringBackup(false);
-    }
-  }, [addToast, closeRestoreOverlay, fetchBackups, restorePreview, restoreSelectionState, restoreSource, restoringBackup]);
-
-  const selectAllRestoreItems = useCallback(() => {
-    if (!restorePreview || !restoreSelectionState) return;
-    setRestoreSelectionState({
-      nodes: Object.fromEntries(restorePreview.config.nodes.map(node => [node.name, true])),
-      checks: Object.fromEntries(restorePreview.config.checks.map(check => [check.id, true])),
-      configFlags: {
-        externalLinks: restorePreview.config.externalLinks.length > 0,
-        registries: restorePreview.config.registries.length > 0,
-        gateway: Boolean(restorePreview.config.gateway),
-        notifications: Boolean(restorePreview.config.notifications),
-        templateSettings: restorePreview.config.templateSettings.length > 0,
-        logLevel: Boolean(restorePreview.config.logLevel),
-        update: Boolean(restorePreview.config.update)
-      },
-      nodeFiles: Object.fromEntries(
-        restorePreview.nodeFiles.map(group => [
-          group.nodeName,
-          Object.fromEntries(group.files.map(file => [file.relativePath, true]))
-        ])
-      ),
-      targetNodes: restoreSelectionState.targetNodes,
-      serviceData: Object.fromEntries((restorePreview.serviceData || []).map(sd => [sd.name, Object.fromEntries(sd.files.map(f => [f, true]))]))
-    });
-  }, [restorePreview, restoreSelectionState]);
-
-  const confirmDeleteBackup = async () => {
-    if (!deleteTarget || deletingBackup) return;
-    setDeletingBackup(true);
-    try {
-      const res = await fetch('/api/settings/backups', {
-        method: 'DELETE',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ fileName: deleteTarget.fileName })
-      });
-      const data = await res.json().catch(() => ({}));
-      if (!res.ok) {
-        throw new Error(data.error || 'Unable to delete backup');
-      }
-      addToast('success', 'Backup deleted', `${deleteTarget.fileName} has been removed.`);
-      await fetchBackups();
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Unknown error';
-      addToast('error', 'Failed to delete backup', message);
-    } finally {
-      setDeletingBackup(false);
-      setDeleteTarget(null);
-    }
-  };
-
-  const availableRestoreTargets = Array.from(new Set(['Local', ...nodes.map(node => node.Name)]));
-
-  const handleRestoreDrop = (event: React.DragEvent<HTMLDivElement>) => {
-    event.preventDefault();
-    if (event.dataTransfer.files && event.dataTransfer.files[0]) {
-      handleRestoreFromFile(event.dataTransfer.files[0]);
-    }
-  };
-
-  const handleRestoreDragOver = (event: React.DragEvent<HTMLDivElement>) => {
-    event.preventDefault();
-  };
-
-  const stopRestoreEvent = useCallback((event: React.MouseEvent) => {
-    event.stopPropagation();
-  }, []);
-
-  const handleRestoreBackdrop = useCallback((event: React.MouseEvent<HTMLDivElement>) => {
-    event.stopPropagation();
-    closeRestoreOverlay();
-  }, [closeRestoreOverlay]);
-
-  const toggleRestoreConfigFlag = (key: keyof BackupRestoreSelection['config']) => {
-    setRestoreSelectionState(prev => {
-      if (!prev) return prev;
-      if (key === 'nodes' || key === 'checks') return prev;
-      return {
-        ...prev,
-        configFlags: {
-          ...prev.configFlags,
-          [key]: !prev.configFlags[key as keyof typeof prev.configFlags]
-        }
-      };
-    });
-  };
-
-  const toggleRestoreNode = (name: string) => {
-    setRestoreSelectionState(prev => {
-      if (!prev) return prev;
-      return {
-        ...prev,
-        nodes: {
-          ...prev.nodes,
-          [name]: !prev.nodes[name]
-        }
-      };
-    });
-  };
-
-  const toggleRestoreCheck = (id: string) => {
-    setRestoreSelectionState(prev => {
-      if (!prev) return prev;
-      return {
-        ...prev,
-        checks: {
-          ...prev.checks,
-          [id]: !prev.checks[id]
-        }
-      };
-    });
-  };
-
-  const toggleRestoreFile = (nodeName: string, filePath: string) => {
-    setRestoreSelectionState(prev => {
-      if (!prev) return prev;
-      return {
-        ...prev,
-        nodeFiles: {
-          ...prev.nodeFiles,
-          [nodeName]: {
-            ...prev.nodeFiles[nodeName],
-            [filePath]: !prev.nodeFiles[nodeName]?.[filePath]
-          }
-        }
-      };
-    });
-  };
-
-  const updateRestoreTargetNode = (sourceNode: string, targetNode: string) => {
-    setRestoreSelectionState(prev => {
-      if (!prev) return prev;
-      return {
-        ...prev,
-        targetNodes: {
-          ...prev.targetNodes,
-          [sourceNode]: targetNode
-        }
-      };
-    });
-  };
-
-  const toggleRestoreSection = (section: string) => {
-    setRestoreExpandedSections(prev => ({ ...prev, [section]: !prev[section] }));
-  };
-
-  const toggleAllNodeFiles = (nodeName: string, selected: boolean) => {
-    if (!restorePreview) return;
-    const group = restorePreview.nodeFiles.find(g => g.nodeName === nodeName);
-    if (!group) return;
-    setRestoreSelectionState(prev => {
-      if (!prev) return prev;
-      return {
-        ...prev,
-        nodeFiles: {
-          ...prev.nodeFiles,
-          [nodeName]: Object.fromEntries(group.files.map(f => [f.relativePath, selected]))
-        }
-      };
-    });
-  };
-
-  const toggleServiceGroupFiles = (nodeName: string, files: { relativePath: string }[], selected: boolean) => {
-    setRestoreSelectionState(prev => {
-      if (!prev) return prev;
-      const updated = { ...prev.nodeFiles[nodeName] };
-      for (const f of files) updated[f.relativePath] = selected;
-      return { ...prev, nodeFiles: { ...prev.nodeFiles, [nodeName]: updated } };
-    });
-  };
-
-  const getRestoreSelectionSummary = () => {
-    if (!restorePreview || !restoreSelectionState) return null;
-    const configCount = Object.values(restoreSelectionState.configFlags).filter(Boolean).length;
-    const nodeCount = Object.values(restoreSelectionState.nodes).filter(Boolean).length;
-    const checkCount = Object.values(restoreSelectionState.checks).filter(Boolean).length;
-    const fileCount = Object.values(restoreSelectionState.nodeFiles).reduce(
-      (sum, files) => sum + Object.values(files).filter(Boolean).length, 0
-    );
-    const dataCount = Object.values(restoreSelectionState.serviceData).reduce(
-      (sum, filesMap) => sum + Object.values(filesMap).filter(Boolean).length, 0
-    );
-    const parts: string[] = [];
-    if (configCount > 0) parts.push(`${configCount} setting${configCount !== 1 ? 's' : ''}`);
-    if (nodeCount > 0) parts.push(`${nodeCount} node${nodeCount !== 1 ? 's' : ''}`);
-    if (checkCount > 0) parts.push(`${checkCount} check${checkCount !== 1 ? 's' : ''}`);
-    if (fileCount > 0) parts.push(`${fileCount} file${fileCount !== 1 ? 's' : ''}`);
-    if (dataCount > 0) parts.push(`${dataCount} data file${dataCount !== 1 ? 's' : ''}`);
-    return parts.length > 0 ? parts.join(', ') : 'Nothing selected';
-  };
-
-  if (loading) {
-    return <div className="p-8 text-center text-gray-500">Loading settings...</div>;
-  }
-
-  return (
-    <div className="h-full overflow-y-auto space-y-6">
-      <PageHeader 
-        title="Settings" 
-        showBack={false}
-        actions={
-          <span className="text-sm text-gray-500 dark:text-gray-400 inline-flex items-center gap-2">
-            {saving ? (
-              <>
-                <Loader2 className="w-4 h-4 animate-spin" />
-                Saving changes…
-              </>
-            ) : (
-              'All changes saved'
-            )}
-          </span>
-        }
-      />
-
-      <div className="flex border-b border-gray-200 dark:border-gray-700 px-6 bg-white dark:bg-gray-900 sticky top-0 z-10">
-        {([
-          { id: 'nodes', label: 'Nodes', icon: Server },
-          { id: 'backups', label: 'Backups', icon: Database },
-          { id: 'notifications', label: 'Notifications', icon: Mail },
-          { id: 'updates', label: 'Updates', icon: RefreshCw },
-          { id: 'advanced', label: 'Advanced', icon: Settings },
-        ] as const).map(tab => (
-          <button
-            key={tab.id}
-            onClick={() => setActiveTab(tab.id)}
-            className={`flex items-center gap-1.5 px-4 py-3 text-sm font-medium border-b-2 transition-colors whitespace-nowrap ${
-              activeTab === tab.id
-                ? 'border-blue-600 text-blue-600 dark:text-blue-400 dark:border-blue-400'
-                : 'border-transparent text-gray-500 hover:text-gray-700 dark:hover:text-gray-300 hover:border-gray-300'
-            }`}
-          >
-            <tab.icon size={16} />
-            {tab.label}
-          </button>
-        ))}
-      </div>
-
-      <div className="px-4 pb-8 w-full space-y-6">
-        {/* ServiceBay Updates */}
-        {activeTab === 'updates' && appUpdate && (
-            <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-                <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex items-center gap-3">
-                  <div className="p-2 bg-purple-100 dark:bg-purple-900/30 rounded-lg text-purple-600 dark:text-purple-400">
-                    <RefreshCw size={20} className={(updateStatus === 'updating') || checkingUpdate ? 'animate-spin' : ''} />
-                  </div>
-                  <div>
-                    <h3 className="font-bold text-gray-900 dark:text-white">ServiceBay Updates</h3>
-                    <p className="text-xs text-gray-500 dark:text-gray-400">Manage application updates</p>
-                  </div>
-                    <div className="ml-auto flex items-center gap-4">
-                        <div className="flex items-center gap-2 text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">Channel:</span>
-                            <select 
-                                value={appUpdate.config.autoUpdate.channel || 'stable'}
-                                onChange={async (e) => {
-                                    const newChannel = e.target.value as 'stable' | 'test' | 'dev';
-                                    try {
-                                        const res = await fetch('/api/system/update', {
-                                            method: 'POST',
-                                            headers: { 'Content-Type': 'application/json' },
-                                            body: JSON.stringify({ 
-                                                action: 'configure', 
-                                                autoUpdate: { 
-                                                    enabled: appUpdate.config.autoUpdate.enabled,
-                                                    channel: newChannel
-                                                } 
-                                            })
-                                        });
-                                        if (res.ok) {
-                                            const data = await res.json();
-                                            setAppUpdate(prev => prev ? { ...prev, config: data.config } : null);
-                                            addToast('success', 'Release channel updated', `Switched to ${newChannel} channel. Check for updates to apply.`);
-                                        }
-                                    } catch {
-                                        addToast('error', 'Failed to update release channel');
-                                    }
-                                }}
-                                className="bg-white dark:bg-gray-800 border border-gray-300 dark:border-gray-600 text-gray-700 dark:text-gray-300 rounded px-2 py-1 text-xs focus:ring-2 focus:ring-purple-500 focus:border-transparent outline-none cursor-pointer"
-                            >
-                                <option value="stable">Stable</option>
-                                <option value="test">Test</option>
-                                <option value="dev">Dev</option>
-                            </select>
-                        </div>
-                        <div className="h-4 w-px bg-gray-300 dark:bg-gray-600"></div>
-                        <button
-                            onClick={handleCheckUpdate}
-                            disabled={checkingUpdate || (updateStatus === 'updating')}
-                            className="text-xs text-purple-600 dark:text-purple-400 hover:underline disabled:opacity-50"
-                        >
-                            {checkingUpdate ? 'Checking...' : 'Check Now'}
-                        </button>
-                        <div className="h-4 w-px bg-gray-300 dark:bg-gray-600"></div>
-                        <label className="relative inline-flex items-center cursor-pointer" title="Enable Auto-Updates">
-                            <input 
-                                type="checkbox" 
-                                className="sr-only peer" 
-                                checked={appUpdate.config.autoUpdate.enabled}
-                                onChange={toggleAutoUpdate}
-                            />
-                            <div className="w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-purple-300 dark:peer-focus:ring-purple-800 rounded-full peer dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-purple-600"></div>
-                        </label>
-                    </div>
-                </div>
-                <div className="p-6">
-                    <div className="flex flex-col md:flex-row gap-6 items-start md:items-center justify-between">
-                        <div className="space-y-1">
-                            <div className="flex items-center gap-2">
-                                <span className="text-sm text-gray-500 dark:text-gray-400">Current Version:</span>
-                                <span className="font-mono font-medium text-gray-900 dark:text-white">{appUpdate.current}</span>
-                            </div>
-                            {appUpdate.hasUpdate && appUpdate.latest ? (
-                                <div className="flex items-center gap-2 text-green-600 dark:text-green-400">
-                                    <Download size={16} />
-                                    <span className="text-sm font-medium">New version available: {appUpdate.latest.version}</span>
-                                </div>
-                            ) : (
-                                <div className="flex items-center gap-2 text-gray-500 dark:text-gray-400">
-                                    <Clock size={16} />
-                                    <span className="text-sm">You are on the latest version</span>
-                                </div>
-                            )}
-                        </div>
-                        
-                        {appUpdate.hasUpdate && appUpdate.latest && (
-                            <button 
-                                onClick={handleAppUpdate}
-                                disabled={updateStatus === 'updating'}
-                                className="px-4 py-2 bg-purple-600 text-white rounded-lg hover:bg-purple-700 transition-colors shadow-sm flex items-center gap-2 disabled:opacity-50 disabled:cursor-not-allowed"
-                            >
-                                <Download size={18} />
-                                {updateStatus === 'updating' ? 'Updating...' : 'Update Now'}
-                            </button>
-                        )}
-                    </div>
-                    
-                    {appUpdate.hasUpdate && appUpdate.latest && appUpdate.latest.notes && (
-                        <div className="mt-4 p-4 bg-gray-50 dark:bg-gray-900/50 rounded-lg border border-gray-200 dark:border-gray-700">
-                            <h4 className="text-sm font-bold text-gray-900 dark:text-white mb-2">Release Notes</h4>
-                            <p className="text-sm text-gray-600 dark:text-gray-300 whitespace-pre-wrap">{appUpdate.latest.notes}</p>
-                        </div>
-                    )}
-                </div>
-            </div>
-        )}
-
-        {/* System Backups */}
-        {activeTab === 'backups' && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-          <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex flex-col gap-3 md:flex-row md:items-center">
-            <div className="flex items-center gap-3 flex-1">
-              <div className="p-2 bg-emerald-100 dark:bg-emerald-900/30 rounded-lg text-emerald-600 dark:text-emerald-300">
-                <HardDrive size={20} />
-              </div>
-              <div>
-                <h3 className="font-bold text-gray-900 dark:text-white">System Backups</h3>
-                <p className="text-xs text-gray-500 dark:text-gray-400">Capture managed services and ServiceBay config into a restorable tarball.</p>
-                {backupStatus !== 'idle' && (
-                  <div className="mt-1 flex items-center gap-2 text-[11px] text-gray-600 dark:text-gray-300">
-                    {backupStatus === 'running' && (
-                      <span className="inline-flex items-center gap-1 text-emerald-600 dark:text-emerald-300">
-                        <Loader2 className="w-3 h-3 animate-spin" />
-                        Backup in progress
-                      </span>
-                    )}
-                    {backupStatus === 'success' && (
-                      <span className="inline-flex items-center gap-1 text-emerald-600 dark:text-emerald-300">
-                        <CheckCircle2 className="w-3 h-3" />
-                        Latest run completed
-                      </span>
-                    )}
-                    {backupStatus === 'error' && (
-                      <span className="inline-flex items-center gap-1 text-red-600 dark:text-red-300">
-                        <XCircle className="w-3 h-3" />
-                        Last run failed
-                      </span>
-                    )}
-                  </div>
-                )}
-              </div>
-            </div>
-            <div className="flex flex-col md:flex-row md:items-center gap-3">
-              <p className="text-[11px] text-gray-500 dark:text-gray-400 bg-white/60 dark:bg-gray-900/40 px-3 py-1 rounded-md border border-gray-200 dark:border-gray-800">
-                Archives stored under <span className="font-mono">~/.config/containers/systemd/backups</span>
-              </p>
-              <button
-                onClick={() => openRestoreOverlay(true)}
-                className="inline-flex items-center justify-center gap-2 px-4 py-2 bg-white dark:bg-gray-900 text-gray-700 dark:text-gray-200 text-sm rounded-lg border border-gray-300 dark:border-gray-700 shadow-sm hover:bg-gray-100 dark:hover:bg-gray-800 transition-colors"
-              >
-                <UploadCloud size={16} />
-                Restore from Backup
-              </button>
-              <button
-                onClick={handleCreateBackup}
-                disabled={creatingBackup}
-                className="inline-flex items-center justify-center gap-2 px-4 py-2 bg-emerald-600 text-white text-sm rounded-lg shadow-sm hover:bg-emerald-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-              >
-                {creatingBackup ? <Loader2 className="animate-spin" size={16} /> : <Download size={16} />}
-                {creatingBackup ? 'Creating Backup...' : 'Create Backup'}
-              </button>
-            </div>
-          </div>
-          <div className="p-6">
-            {backupsLoading ? (
-              <div className="flex items-center gap-3 text-sm text-gray-500 dark:text-gray-400">
-                <Loader2 className="animate-spin" size={18} />
-                Loading backups...
-              </div>
-            ) : backups.length === 0 ? (
-              <div className="text-sm text-gray-500 dark:text-gray-400 italic">No backups found. Create one to snapshot your environment.</div>
-            ) : (
-              <div className="overflow-x-auto">
-                <table className="min-w-full text-sm">
-                  <thead>
-                    <tr className="text-left text-gray-500 dark:text-gray-400 border-b border-gray-200 dark:border-gray-800">
-                      <th className="py-2 font-medium">Archive</th>
-                      <th className="py-2 font-medium">Created</th>
-                      <th className="py-2 font-medium">Size</th>
-                      <th className="py-2 font-medium text-right">Actions</th>
-                    </tr>
-                  </thead>
-                  <tbody className="divide-y divide-gray-200 dark:divide-gray-800">
-                    {backups.map(backup => (
-                      <tr key={backup.fileName}>
-                        <td className="py-3 font-mono text-xs text-blue-600 dark:text-blue-300 break-all">{backup.fileName}</td>
-                        <td className="py-3 text-gray-700 dark:text-gray-300">{new Date(backup.createdAt).toLocaleString()}</td>
-                        <td className="py-3 text-gray-700 dark:text-gray-300">{formatBytes(backup.size)}</td>
-                        <td className="py-3">
-                          <div className="flex justify-end gap-2">
-                            <button
-                              onClick={() => handleDownloadBackup(backup.fileName)}
-                              className="text-xs px-3 py-1.5 rounded-md border border-gray-300 dark:border-gray-700 text-gray-700 dark:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-800 transition-colors flex items-center gap-1"
-                            >
-                              <Download size={14} />
-                              Download
-                            </button>
-                            <button
-                              onClick={() => handleRestoreRequest(backup)}
-                              className="text-xs px-3 py-1.5 rounded-md border border-amber-300 text-amber-700 dark:text-amber-300 dark:border-amber-700 hover:bg-amber-50 dark:hover:bg-amber-900/20 transition-colors flex items-center gap-1"
-                            >
-                              <RotateCcw size={14} />
-                              Restore
-                            </button>
-                                <button
-                                  onClick={() => setDeleteTarget(backup)}
-                                  disabled={deletingBackup}
-                                  className="text-xs px-3 py-1.5 rounded-md border border-red-200 text-red-600 dark:text-red-400 dark:border-red-700 hover:bg-red-50 dark:hover:bg-red-900/20 transition-colors flex items-center gap-1 disabled:opacity-60"
-                                >
-                                  <Trash2 size={14} />
-                                  Delete
-                                </button>
-                          </div>
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            )}
-
-                {(backupLog.length > 0 || backupStatus === 'running' || backupStatus === 'error') && (
-                  <div className="mt-6 border border-gray-200 dark:border-gray-800 rounded-lg p-4 bg-gray-50/60 dark:bg-gray-900/40">
-                    <div className="flex items-center justify-between mb-3">
-                      <span className="text-sm font-semibold text-gray-800 dark:text-gray-100">Backup Activity</span>
-                      {backupStatus === 'running' && (
-                        <span className="inline-flex items-center gap-1 text-xs text-emerald-600 dark:text-emerald-300">
-                          <Loader2 className="w-3 h-3 animate-spin" />
-                          Streaming logs
-                        </span>
-                      )}
-                      {backupStatus === 'error' && (
-                        <span className="inline-flex items-center gap-1 text-xs text-red-600 dark:text-red-300">
-                          <XCircle className="w-3 h-3" />
-                          Check details below
-                        </span>
-                      )}
-                      {backupStatus === 'success' && backupLog.length > 0 && (
-                        <span className="inline-flex items-center gap-1 text-xs text-emerald-600 dark:text-emerald-300">
-                          <CheckCircle2 className="w-3 h-3" />
-                          Completed
-                        </span>
-                      )}
-                    </div>
-                    <div className="max-h-48 overflow-y-auto pr-1 space-y-3">
-                      {backupLog.length === 0 ? (
-                        <p className="text-xs text-gray-500 dark:text-gray-400 italic">Waiting for backup updates…</p>
-                      ) : (
-                        backupLog.map((entry, idx) => (
-                          <div key={`${entry.timestamp}-${idx}`} className="flex gap-3 text-xs">
-                            <span className={`mt-1 h-2 w-2 rounded-full ${LOG_STATUS_DOTS[entry.status] ?? LOG_STATUS_DOTS.info}`}></span>
-                            <div className="flex-1">
-                              <div className="flex flex-wrap items-center gap-2 text-[11px] text-gray-500 dark:text-gray-400">
-                                <span className="font-mono">{new Date(entry.timestamp).toLocaleTimeString()}</span>
-                                {entry.node && <span className="uppercase tracking-wide text-gray-600 dark:text-gray-300">{entry.node}</span>}
-                                <span className={`px-2 py-0.5 rounded ${LOG_STATUS_BADGES[entry.status] ?? LOG_STATUS_BADGES.info}`}>
-                                  {entry.status.toUpperCase()}
-                                </span>
-                              </div>
-                              <p className="text-gray-700 dark:text-gray-200">{entry.message}</p>
-                              {entry.target && (
-                                <p className="text-[10px] font-mono text-gray-500 dark:text-gray-400 break-all">{entry.target}</p>
-                              )}
-                            </div>
-                          </div>
-                        ))
-                      )}
-                    </div>
-                  </div>
-                )}
-          </div>
-        </div>
-        )}
-
-        {/* Backup Sync (RAID → External Target) */}
-        {activeTab === 'backups' && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-          <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50">
-            <div className="flex items-center gap-3">
-              <div className="p-2 bg-blue-100 dark:bg-blue-900/30 rounded-lg text-blue-600 dark:text-blue-300">
-                <UploadCloud size={20} />
-              </div>
-              <div className="flex-1">
-                <h3 className="font-bold text-gray-900 dark:text-white">Backup Sync</h3>
-                <p className="text-xs text-gray-500 dark:text-gray-400">
-                  Incrementally sync your data to an external target using rsync.
-                </p>
-              </div>
-              <label className="flex items-center gap-2 cursor-pointer">
-                <span className="text-xs text-gray-500 dark:text-gray-400">{backupSync.enabled ? 'Enabled' : 'Disabled'}</span>
-                <div className="relative">
-                  <input type="checkbox" className="sr-only peer" checked={backupSync.enabled} onChange={e => setBackupSync(prev => ({ ...prev, enabled: e.target.checked }))} />
-                  <div className="w-9 h-5 bg-gray-200 peer-focus:outline-none rounded-full peer dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:start-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-4 after:w-4 after:transition-all peer-checked:bg-blue-600"></div>
-                </div>
-              </label>
-            </div>
-            {backupSync.lastRun && (
-              <div className="mt-2 text-xs text-gray-500 dark:text-gray-400 flex items-center gap-2">
-                <Clock size={12} />
-                Last run: {new Date(backupSync.lastRun).toLocaleString()}
-                {backupSync.lastStatus === 'success' && <span className="text-emerald-600"><CheckCircle2 size={12} className="inline" /></span>}
-                {backupSync.lastStatus === 'error' && <span className="text-red-500"><XCircle size={12} className="inline" /></span>}
-                {backupSync.lastDuration != null && <span>({backupSync.lastDuration}s)</span>}
-              </div>
-            )}
-          </div>
-
-          <div className="p-4 space-y-4">
-            {/* Source */}
-            <div>
-              <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Source Path</label>
-              <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.sourcePath} onChange={e => setBackupSync(prev => ({ ...prev, sourcePath: e.target.value }))} placeholder="/mnt/data" />
-            </div>
-
-            {/* Target Type */}
-            <div>
-              <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Target Type</label>
-              <div className="grid grid-cols-4 gap-2">
-                {([['local', 'Local / USB'], ['ssh', 'SSH'], ['smb', 'SMB / CIFS'], ['nfs', 'NFS']] as const).map(([val, label]) => (
-                  <button key={val} onClick={() => setBackupSync(prev => ({ ...prev, targetType: val }))}
-                    className={`px-3 py-2 text-xs rounded-lg border transition-colors ${backupSync.targetType === val ? 'bg-blue-50 dark:bg-blue-900/30 border-blue-300 dark:border-blue-600 text-blue-700 dark:text-blue-300 font-medium' : 'border-gray-300 dark:border-gray-600 text-gray-600 dark:text-gray-400 hover:bg-gray-50 dark:hover:bg-gray-700'}`}>
-                    {label}
-                  </button>
-                ))}
-              </div>
-            </div>
-
-            {/* Target fields */}
-            {backupSync.targetType === 'local' && (
-              <div>
-                <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Target Path</label>
-                <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.localPath} onChange={e => setBackupSync(prev => ({ ...prev, localPath: e.target.value }))} placeholder="/mnt/backup" />
-                <p className="text-[11px] text-gray-400 mt-1">Mount your USB/external drive here first.</p>
-              </div>
-            )}
-
-            {backupSync.targetType === 'ssh' && (
-              <div className="grid grid-cols-2 gap-3">
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Host</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.sshHost} onChange={e => setBackupSync(prev => ({ ...prev, sshHost: e.target.value }))} placeholder="192.168.1.100" />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Port</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.sshPort} onChange={e => setBackupSync(prev => ({ ...prev, sshPort: e.target.value }))} placeholder="22" />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">User</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.sshUser} onChange={e => setBackupSync(prev => ({ ...prev, sshUser: e.target.value }))} />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Remote Path</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.sshPath} onChange={e => setBackupSync(prev => ({ ...prev, sshPath: e.target.value }))} placeholder="/backup" />
-                </div>
-                <div className="col-span-2">
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Identity File</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.sshIdentityFile} onChange={e => setBackupSync(prev => ({ ...prev, sshIdentityFile: e.target.value }))} />
-                </div>
-              </div>
-            )}
-
-            {backupSync.targetType === 'smb' && (
-              <div className="grid grid-cols-2 gap-3">
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Host</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.smbHost} onChange={e => setBackupSync(prev => ({ ...prev, smbHost: e.target.value }))} placeholder="nas.local" />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Share Name</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.smbShare} onChange={e => setBackupSync(prev => ({ ...prev, smbShare: e.target.value }))} placeholder="backup" />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Subfolder (optional)</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.smbPath} onChange={e => setBackupSync(prev => ({ ...prev, smbPath: e.target.value }))} />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Username</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.smbUsername} onChange={e => setBackupSync(prev => ({ ...prev, smbUsername: e.target.value }))} />
-                </div>
-                <div className="col-span-2">
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Password</label>
-                  <input type="password" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.smbPassword} onChange={e => setBackupSync(prev => ({ ...prev, smbPassword: e.target.value }))} />
-                </div>
-              </div>
-            )}
-
-            {backupSync.targetType === 'nfs' && (
-              <div className="grid grid-cols-2 gap-3">
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Host</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.nfsHost} onChange={e => setBackupSync(prev => ({ ...prev, nfsHost: e.target.value }))} placeholder="nas.local" />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Export Path</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.nfsExport} onChange={e => setBackupSync(prev => ({ ...prev, nfsExport: e.target.value }))} placeholder="/volume1/backup" />
-                </div>
-                <div className="col-span-2">
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Subfolder (optional)</label>
-                  <input type="text" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.nfsPath} onChange={e => setBackupSync(prev => ({ ...prev, nfsPath: e.target.value }))} />
-                </div>
-              </div>
-            )}
-
-            {/* Schedule */}
-            <div className="grid grid-cols-2 gap-3">
-              <div>
-                <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Schedule</label>
-                <select className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.schedule} onChange={e => setBackupSync(prev => ({ ...prev, schedule: e.target.value as 'hourly' | 'daily' | 'weekly' | 'monthly' }))}>
-                  <option value="hourly">Hourly</option>
-                  <option value="daily">Daily</option>
-                  <option value="weekly">Weekly</option>
-                  <option value="monthly">Monthly</option>
-                </select>
-              </div>
-              <div>
-                <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Time (UTC)</label>
-                <input type="time" className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.time} onChange={e => setBackupSync(prev => ({ ...prev, time: e.target.value }))} />
-              </div>
-              {backupSync.schedule === 'weekly' && (
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Day of Week</label>
-                  <select className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.dayOfWeek ?? 0} onChange={e => setBackupSync(prev => ({ ...prev, dayOfWeek: parseInt(e.target.value) }))}>
-                    {['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'].map((d, i) => (
-                      <option key={i} value={i}>{d}</option>
-                    ))}
-                  </select>
-                </div>
-              )}
-              {backupSync.schedule === 'monthly' && (
-                <div>
-                  <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Day of Month</label>
-                  <input type="number" min={1} max={28} className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white" value={backupSync.dayOfMonth ?? 1} onChange={e => setBackupSync(prev => ({ ...prev, dayOfMonth: parseInt(e.target.value) || 1 }))} />
-                </div>
-              )}
-            </div>
-
-            {/* Exclude Patterns */}
-            <div>
-              <label className="block text-xs font-medium text-gray-700 dark:text-gray-300 mb-1">Exclude Patterns (one per line)</label>
-              <textarea className="w-full px-3 py-2 text-sm rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white font-mono" rows={2} value={backupSync.excludePatterns} onChange={e => setBackupSync(prev => ({ ...prev, excludePatterns: e.target.value }))} placeholder="*.tmp&#10;*.log" />
-            </div>
-
-            {/* Test Result */}
-            {backupSyncTestResult && (
-              <div className={`p-3 text-sm rounded-lg ${backupSyncTestResult.success ? 'bg-emerald-50 dark:bg-emerald-900/20 text-emerald-700 dark:text-emerald-300' : 'bg-red-50 dark:bg-red-900/20 text-red-700 dark:text-red-300'}`}>
-                {backupSyncTestResult.success ? <CheckCircle2 size={14} className="inline mr-1" /> : <XCircle size={14} className="inline mr-1" />}
-                {backupSyncTestResult.message}
-              </div>
-            )}
-
-            {/* Actions */}
-            <div className="flex flex-wrap gap-2 pt-2 border-t border-gray-200 dark:border-gray-700">
-              <button onClick={handleSaveBackupSync} disabled={backupSyncSaving}
-                className="flex items-center gap-2 px-4 py-2 text-sm font-medium text-white bg-blue-600 rounded-lg hover:bg-blue-700 disabled:opacity-50">
-                {backupSyncSaving ? <Loader2 size={14} className="animate-spin" /> : <Save size={14} />}
-                Save
-              </button>
-              <button onClick={handleTestBackupSync} disabled={backupSyncTesting}
-                className="flex items-center gap-2 px-4 py-2 text-sm font-medium text-gray-700 dark:text-gray-200 bg-gray-100 dark:bg-gray-700 rounded-lg hover:bg-gray-200 dark:hover:bg-gray-600 disabled:opacity-50">
-                {backupSyncTesting ? <Loader2 size={14} className="animate-spin" /> : <Activity size={14} />}
-                Test Connection
-              </button>
-              <button onClick={handleRunBackupSync} disabled={backupSyncRunning}
-                className="flex items-center gap-2 px-4 py-2 text-sm font-medium text-white bg-emerald-600 rounded-lg hover:bg-emerald-700 disabled:opacity-50">
-                {backupSyncRunning ? <Loader2 size={14} className="animate-spin" /> : <RefreshCw size={14} />}
-                {backupSyncRunning ? 'Running...' : 'Run Now'}
-              </button>
-            </div>
-
-            {/* History */}
-            {backupSyncHistory.length > 0 && (
-              <div className="pt-3 border-t border-gray-200 dark:border-gray-700">
-                <h4 className="text-xs font-medium text-gray-700 dark:text-gray-300 mb-2">Recent Runs</h4>
-                <div className="space-y-1 max-h-40 overflow-y-auto">
-                  {backupSyncHistory.slice(0, 10).map((h, i) => (
-                    <div key={i} className="flex items-center gap-2 text-xs text-gray-500 dark:text-gray-400">
-                      {h.success ? <CheckCircle2 size={12} className="text-emerald-500 flex-shrink-0" /> : <XCircle size={12} className="text-red-500 flex-shrink-0" />}
-                      <span className="font-mono">{new Date(h.startedAt).toLocaleString()}</span>
-                      <span>({h.duration}s)</span>
-                      {h.filesTransferred != null && <span>{h.filesTransferred} files</span>}
-                      <span className="truncate flex-1">{h.message}</span>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-        )}
-
-        {/* Nginx Config Export/Import */}
-        {activeTab === 'backups' && nginxInstalled && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-          <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex flex-col gap-3 md:flex-row md:items-center">
-            <div className="flex items-center gap-3 flex-1">
-              <div className="p-2 bg-green-100 dark:bg-green-900/30 rounded-lg text-green-600 dark:text-green-300">
-                <Shield size={20} />
-              </div>
-              <div>
-                <h3 className="font-bold text-gray-900 dark:text-white">Nginx Configuration</h3>
-                <p className="text-xs text-gray-500 dark:text-gray-400">
-                  Export or import reverse proxy server block configs{nginxNode && nginxNode !== 'Local' ? ` (${nginxNode})` : ''}.
-                </p>
-              </div>
-            </div>
-            <div className="flex flex-col md:flex-row md:items-center gap-3">
-              <button
-                onClick={handleNginxExport}
-                disabled={nginxExporting}
-                className="inline-flex items-center justify-center gap-2 px-4 py-2 bg-white dark:bg-gray-900 text-gray-700 dark:text-gray-200 text-sm rounded-lg border border-gray-300 dark:border-gray-700 shadow-sm hover:bg-gray-100 dark:hover:bg-gray-800 transition-colors disabled:opacity-50"
-              >
-                {nginxExporting ? <Loader2 className="animate-spin" size={16} /> : <Download size={16} />}
-                Export Config
-              </button>
-              <label className="inline-flex items-center justify-center gap-2 px-4 py-2 bg-white dark:bg-gray-900 text-gray-700 dark:text-gray-200 text-sm rounded-lg border border-gray-300 dark:border-gray-700 shadow-sm hover:bg-gray-100 dark:hover:bg-gray-800 transition-colors cursor-pointer">
-                {nginxImporting ? <Loader2 className="animate-spin" size={16} /> : <Upload size={16} />}
-                Import Config
-                <input
-                  ref={nginxFileInputRef}
-                  type="file"
-                  accept=".json,.tar.gz,.tgz"
-                  onChange={handleNginxImport}
-                  className="hidden"
-                />
-              </label>
-            </div>
-          </div>
-          {nginxDiag && (
-            <div className="p-4 border-t border-red-200 dark:border-red-800/50 bg-red-50 dark:bg-red-900/20">
-              <div className="flex items-start gap-3">
-                <XCircle className="w-5 h-5 text-red-500 dark:text-red-400 shrink-0 mt-0.5" />
-                <div className="flex-1 min-w-0">
-                  <p className="text-sm text-red-800 dark:text-red-200">{nginxDiag.reason}</p>
-                  {(nginxDiag.node || nginxDiag.confDir) && (
-                    <div className="mt-1 flex flex-wrap gap-x-4 gap-y-1 text-xs text-red-600 dark:text-red-300">
-                      {nginxDiag.node && <span>Node: <span className="font-mono">{nginxDiag.node}</span></span>}
-                      {nginxDiag.confDir && <span>conf.d: <span className="font-mono">{nginxDiag.confDir}</span></span>}
-                    </div>
-                  )}
-                  {nginxDiag.debug.length > 0 && (
-                    <div className="mt-2">
-                      <button
-                        onClick={() => setNginxDiagExpanded(prev => !prev)}
-                        className="inline-flex items-center gap-1 text-xs text-red-600 dark:text-red-300 hover:text-red-800 dark:hover:text-red-100 transition-colors"
-                      >
-                        {nginxDiagExpanded ? <ChevronDown size={14} /> : <ChevronRight size={14} />}
-                        Diagnostics ({nginxDiag.debug.length} steps)
-                      </button>
-                      {nginxDiagExpanded && (
-                        <pre className="mt-1 p-2 rounded bg-red-100 dark:bg-red-900/40 text-[11px] font-mono text-red-700 dark:text-red-200 overflow-x-auto max-h-60 overflow-y-auto whitespace-pre-wrap break-words">
-                          {nginxDiag.debug.join('\n')}
-                        </pre>
-                      )}
-                    </div>
-                  )}
-                </div>
-                <button onClick={() => setNginxDiag(null)} className="text-red-400 hover:text-red-600 dark:hover:text-red-200 transition-colors shrink-0">
-                  <X size={16} />
-                </button>
-              </div>
-            </div>
-          )}
-        </div>
-        )}
-
-        {/* Server Identity */}
-        {activeTab === 'advanced' && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-          <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex items-center gap-3">
-            <div className="p-2 bg-indigo-100 dark:bg-indigo-900/30 rounded-lg text-indigo-600 dark:text-indigo-400">
-              <Server size={20} />
-            </div>
-            <div>
-              <h3 className="font-bold text-gray-900 dark:text-white">Server Identity</h3>
-              <p className="text-xs text-gray-500 dark:text-gray-400">Custom display name shown in the browser tab and system info instead of the detected hostname.</p>
-            </div>
-          </div>
-          <div className="p-6">
-            <div className="flex items-center gap-3">
-              <input
-                type="text"
-                value={serverName}
-                onChange={e => setServerName(e.target.value)}
-                disabled={saving}
-                className="flex-1 p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-indigo-500 outline-none"
-                placeholder="e.g. HomeServer, NAS, Production"
-              />
-              <button
-                onClick={() => persistSettings()}
-                disabled={saving}
-                className="px-4 py-2 bg-indigo-600 text-white rounded-lg hover:bg-indigo-700 disabled:opacity-50 transition-colors text-sm font-medium"
-              >
-                {saving ? 'Saving...' : 'Save'}
-              </button>
-            </div>
-          </div>
-        </div>
-        )}
-
-        {/* Template Settings Section */}
-        {activeTab === 'advanced' && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-          <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex items-center gap-3">
-            <div className="p-2 bg-indigo-100 dark:bg-indigo-900/30 rounded-lg text-indigo-600 dark:text-indigo-400">
-              <Server size={20} />
-            </div>
-            <div>
-              <h3 className="font-bold text-gray-900 dark:text-white">Template Settings</h3>
-              <p className="text-xs text-gray-500 dark:text-gray-400">Define global variables used when rendering new stacks (e.g., DATA_DIR). Updates affect future deployments only.</p>
-            </div>
-          </div>
-          <div className="p-6 space-y-6">
-            <div className="space-y-4">
-              <div className="p-4 border border-gray-200 dark:border-gray-700 rounded-lg bg-gradient-to-r from-indigo-50 via-white to-white dark:from-gray-900 dark:via-gray-900 dark:to-gray-900 flex flex-col md:flex-row md:items-center md:gap-4">
-                <div className="flex-1">
-                  <p className="text-sm font-semibold text-gray-900 dark:text-white mb-2">Add Variable</p>
-                  <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
-                    <input
-                      type="text"
-                      value={newVarKey}
-                      onChange={e => setNewVarKey(e.target.value)}
-                      disabled={saving}
-                      className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-indigo-500 outline-none"
-                      placeholder="VAR_NAME"
-                    />
-                    <input
-                      type="text"
-                      value={newVarValue}
-                      onChange={e => setNewVarValue(e.target.value)}
-                      disabled={saving}
-                      className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-indigo-500 outline-none"
-                      placeholder="value"
-                    />
-                  </div>
-                  <p className="text-xs text-gray-500 dark:text-gray-400 mt-2">Variables appear below immediately after you add them.</p>
-                </div>
-                <button
-                  onClick={handleAddTemplateVariable}
-                  disabled={saving || !newVarKey.trim()}
-                  className="w-full md:w-auto px-4 py-2 bg-indigo-600 text-white rounded-lg hover:bg-indigo-700 transition-colors shadow-sm flex items-center gap-2 justify-center disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  <Plus size={16} />
-                  Add Variable
-                </button>
-              </div>
-
-              {Object.keys(templateValues).sort().map(key => {
-                const meta = templateSchema[key];
-                const isRequired = meta?.required;
-                return (
-                  <div key={key} className="p-4 border border-gray-200 dark:border-gray-700 rounded-lg bg-gray-50 dark:bg-gray-900 flex flex-col md:flex-row md:items-center md:gap-4">
-                    <div className="flex-1">
-                      <div className="flex items-center gap-2 mb-1">
-                        <span className="font-semibold text-gray-900 dark:text-white">{key}</span>
-                        {isRequired && (
-                          <span className="text-xs px-2 py-0.5 rounded bg-indigo-100 text-indigo-700 dark:bg-indigo-900/40 dark:text-indigo-200">Required</span>
-                        )}
-                      </div>
-                      <p className="text-xs text-gray-500 dark:text-gray-400 mb-2">
-                        {meta?.description || 'Template variable'}
-                        {meta?.default ? ` (default: ${meta.default})` : ''}
-                      </p>
-                      <input
-                        type="text"
-                        value={templateValues[key] || ''}
-                        onChange={e => handleTemplateValueChange(key, e.target.value)}
-                        onBlur={() => persistSettings()}
-                        disabled={saving}
-                        className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-indigo-500 outline-none"
-                        placeholder={meta?.default || ''}
-                      />
-                    </div>
-                    {!isRequired && (
-                      <button
-                        onClick={() => handleRemoveTemplateVariable(key)}
-                        disabled={saving}
-                        className="mt-3 md:mt-0 text-gray-400 hover:text-red-500 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                        aria-label={`Remove ${key}`}
-                      >
-                        <Trash2 size={18} />
-                      </button>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-
-          </div>
-        </div>
-        )}
-
-
-  {/* System Connections (Nodes) */}
-        {activeTab === 'nodes' && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-            <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex items-center gap-3">
-                <div className="p-2 bg-blue-100 dark:bg-blue-900/30 rounded-lg text-blue-600 dark:text-blue-400">
-                    <Server size={20} />
-                </div>
-                <div>
-                    <h3 className="font-bold text-gray-900 dark:text-white">System Connections</h3>
-                    <p className="text-xs text-gray-500 dark:text-gray-400">Manage remote Podman nodes</p>
-                </div>
-                <div className="ml-auto">
-                    <button 
-                        onClick={() => setIsSSHModalOpen(true)}
-                        className="text-xs flex items-center gap-1 text-blue-600 dark:text-blue-400 hover:underline"
-                    >
-                        <Terminal size={14} />
-                        Setup SSH Keys
-                    </button>
-                </div>
-            </div>
-            
-            <div className="p-6 space-y-6">
-                <div className="bg-blue-50 dark:bg-blue-900/20 border border-blue-200 dark:border-blue-800 rounded-lg p-3 flex gap-3 items-start">
-                    <div className="mt-0.5 text-blue-600 dark:text-blue-400">
-                        <Key size={16} />
-                    </div>
-                    <div className="text-sm text-blue-800 dark:text-blue-200">
-                        <p className="font-medium mb-1">SSH Access Required</p>
-                        <p className="opacity-90 text-xs">
-                            ServiceBay requires password-less SSH access to remote nodes. 
-                            If you haven&apos;t set this up, use the 
-                            <button onClick={() => setIsSSHModalOpen(true)} className="mx-1 underline font-medium hover:text-blue-600">Setup SSH Keys</button> 
-                            tool to copy your public key to the server.
-                        </p>
-                    </div>
-                </div>
-
-                <div className="grid grid-cols-1 md:grid-cols-12 gap-4 items-end mb-6" id="node-form">
-                    <div className="md:col-span-3">
-                        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Name</label>
-                        <input 
-                            type="text" 
-                            value={newNodeName}
-                        onChange={e => setNewNodeName(e.target.value)}
-                        disabled={addingNode}
-                            className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none disabled:opacity-50"
-                            placeholder="my-node"
-                        />
-                    </div>
-                    <div className="md:col-span-5">
-                        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Destination (SSH)</label>
-                        <input 
-                            type="text" 
-                            value={newNodeDest}
-                        onChange={e => setNewNodeDest(e.target.value)}
-                            disabled={addingNode}
-                            className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none disabled:opacity-50 disabled:cursor-not-allowed"
-                            placeholder="ssh://user@host:port"
-                        />
-                    </div>
-                    <div className="md:col-span-3">
-                        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Identity File</label>
-                        <div className="relative">
-                            <Key className="absolute left-3 top-1/2 -translate-y-1/2 w-4 h-4 text-gray-400" />
-                            <input 
-                                type="text" 
-                                value={newNodeIdentity}
-                              onChange={e => setNewNodeIdentity(e.target.value)}
-                                disabled={addingNode}
-                                className="w-full pl-9 pr-2 py-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none disabled:opacity-50 disabled:cursor-not-allowed"
-                                placeholder="/app/data/ssh/id_rsa"
-                            />
-                        </div>
-                    </div>
-                    <div className="md:col-span-1 flex gap-2">
-                        <button 
-                            onClick={handleAddNode}
-                            disabled={addingNode || !newNodeName.trim() || !newNodeDest.trim() || !newNodeIdentity.trim()}
-                            className={`w-full p-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex justify-center items-center gap-2`}
-                            title="Add Node"
-                        >
-                            {addingNode ? <Loader2 className="animate-spin" size={20} /> : <Plus size={20} />}
-                        </button>
-                    </div>
-                </div>
-
-                <div className="space-y-2">
-                  {nodes.map(node => {
-                    const health = nodeHealth[node.Name] || { loading: false, online: false, auth: false };
-                    const isEditing = editingNode === node.Name;
-                    const displayName = isEditing ? (nodeDraft.name || node.Name) : node.Name;
-                    const inlineDisabled = savingNode || !nodeDraft.name.trim() || !nodeDraft.destination.trim() || !nodeDraft.identity.trim();
-
-                    return (
-                    <div key={node.Name} className={`flex flex-col gap-4 md:flex-row md:items-start md:justify-between p-4 rounded-lg border transition-colors ${isEditing ? 'bg-blue-50 border-blue-200 dark:bg-blue-900/20 dark:border-blue-800' : 'bg-gray-50 dark:bg-gray-900/50 border-gray-200 dark:border-gray-700'}`}>
-                      <div className="flex-1 space-y-3">
-                        <div className="flex items-center gap-3">
-                          <div className={`w-2 h-2 rounded-full ${node.Default ? 'bg-green-500' : 'bg-gray-300 dark:bg-gray-600'}`} title={node.Default ? 'Default Node' : ''} />
-                          <div>
-                            <div className="font-medium text-gray-900 dark:text-white flex flex-wrap items-center gap-2">
-                              {displayName}
-                              {node.Default && <span className="text-[10px] bg-green-100 dark:bg-green-900/30 text-green-600 dark:text-green-400 px-1.5 py-0.5 rounded uppercase font-bold">Default</span>}
-                              <div className="flex items-center gap-1 ml-1" title={health.error || (health.online ? (health.auth ? 'Online & Authenticated' : 'Online but Auth Failed') : 'Unreachable')}>
-                                {health.loading ? (
-                                  <Loader2 size={14} className="animate-spin text-gray-400" />
-                                ) : health.online && health.auth ? (
-                                  <div className="flex items-center text-green-500 gap-1 text-[10px] bg-white dark:bg-gray-800 px-1.5 py-0.5 rounded border border-gray-200 dark:border-gray-700 shadow-sm">
-                                    <Globe size={10} />
-                                    <span>Connected</span>
-                                  </div>
-                                ) : health.online && !health.auth ? (
-                                  <div className="flex items-center text-yellow-500 gap-1 text-[10px] bg-white dark:bg-gray-800 px-1.5 py-0.5 rounded border border-gray-200 dark:border-gray-700 shadow-sm cursor-help">
-                                    <ShieldAlert size={10} />
-                                    <span>Auth Failed</span>
-                                  </div>
-                                ) : (
-                                  <div className="flex items-center text-red-500 gap-1 text-[10px] bg-white dark:bg-gray-800 px-1.5 py-0.5 rounded border border-gray-200 dark:border-gray-700 shadow-sm cursor-help">
-                                    <WifiOff size={10} />
-                                    <span>Offline</span>
-                                  </div>
-                                )}
-                              </div>
-                            </div>
-                            {!isEditing && <div className="text-xs text-gray-500 dark:text-gray-400 font-mono">{node.URI}</div>}
-                          </div>
-                        </div>
-
-                        {isEditing && (
-                          <div className="grid grid-cols-1 md:grid-cols-12 gap-3">
-                            <div className="md:col-span-3">
-                              <label className="block text-xs font-medium text-gray-600 dark:text-gray-300 mb-1">Name</label>
-                              <input
-                                type="text"
-                                value={nodeDraft.name}
-                                onChange={e => setNodeDraft(prev => ({ ...prev, name: e.target.value }))}
-                                disabled={savingNode}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white text-sm focus:ring-2 focus:ring-blue-500 outline-none disabled:opacity-50 disabled:cursor-not-allowed"
-                                placeholder="my-node"
-                              />
-                            </div>
-                            <div className="md:col-span-5">
-                              <label className="block text-xs font-medium text-gray-600 dark:text-gray-300 mb-1">Destination (SSH)</label>
-                              <input
-                                type="text"
-                                value={nodeDraft.destination}
-                                onChange={e => setNodeDraft(prev => ({ ...prev, destination: e.target.value }))}
-                                disabled={savingNode}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white text-sm focus:ring-2 focus:ring-blue-500 outline-none disabled:opacity-50 disabled:cursor-not-allowed"
-                                placeholder="ssh://user@host:port"
-                              />
-                            </div>
-                            <div className="md:col-span-4">
-                              <label className="block text-xs font-medium text-gray-600 dark:text-gray-300 mb-1">Identity File</label>
-                              <input
-                                type="text"
-                                value={nodeDraft.identity}
-                                onChange={e => setNodeDraft(prev => ({ ...prev, identity: e.target.value }))}
-                                disabled={savingNode}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white text-sm focus:ring-2 focus:ring-blue-500 outline-none disabled:opacity-50 disabled:cursor-not-allowed"
-                                placeholder="/app/data/ssh/id_rsa"
-                              />
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                      <div className="flex items-center gap-2">
-                        {isEditing ? (
-                          <>
-                            <button
-                              onClick={handleInlineSave}
-                              disabled={inlineDisabled}
-                              className="p-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-                              title="Save changes"
-                            >
-                              {savingNode ? <Loader2 size={18} className="animate-spin" /> : <Save size={18} />}
-                            </button>
-                            <button
-                              onClick={cancelInlineEdit}
-                              className="p-2 bg-gray-100 dark:bg-gray-800 text-gray-700 dark:text-gray-300 rounded-lg hover:bg-gray-200 dark:hover:bg-gray-700 transition-colors"
-                              title="Cancel"
-                            >
-                              <XCircle size={18} />
-                            </button>
-                          </>
-                        ) : (
-                          <>
-                            {!node.Default && (
-                              <button 
-                                onClick={() => handleSetDefaultNode(node.Name)}
-                                className="p-2 text-gray-500 hover:text-green-600 hover:bg-green-50 dark:hover:bg-green-900/20 rounded transition-colors"
-                                title="Set as Default"
-                              >
-                                <CheckCircle2 size={16} />
-                              </button>
-                            )}
-                            <button
-                              onClick={() => handleDeploySSHKey(node)}
-                              className="p-2 text-gray-500 hover:text-indigo-600 hover:bg-indigo-50 dark:hover:bg-indigo-900/20 rounded transition-colors"
-                              title="Deploy SSH key to this node"
-                            >
-                              <Key size={16} />
-                            </button>
-                            <button
-                              onClick={() => handleOpenTerminal(node)}
-                              className="p-2 text-gray-500 hover:text-purple-600 hover:bg-purple-50 dark:hover:bg-purple-900/20 rounded transition-colors"
-                              title="Open SSH terminal for this node"
-                            >
-                              <Terminal size={16} />
-                            </button>
-                            <button 
-                              onClick={() => startEditingNode(node)}
-                              className="p-2 text-gray-500 hover:text-blue-600 hover:bg-blue-50 dark:hover:bg-blue-900/20 rounded transition-colors"
-                              title="Edit Node settings"
-                            >
-                              <Edit2 size={16} />
-                            </button>
-                            <button 
-                              onClick={() => handleDeleteNode(node.Name)}
-                              className="p-2 text-gray-500 hover:text-red-600 hover:bg-red-50 dark:hover:bg-red-900/20 rounded transition-colors"
-                              title="Remove Node"
-                            >
-                              <Trash2 size={16} />
-                            </button>
-                          </>
-                        )}
-                      </div>
-                    </div>
-                  );
-                  })}
-                    {nodes.length === 0 && (
-                        <div className="text-center py-4 text-gray-500 dark:text-gray-400 text-sm italic">
-                            No remote nodes configured. ServiceBay is running in local mode.
-                        </div>
-                    )}
-                </div>
-            </div>
-        </div>
-        )}
-
-        {/* Template Registries */}
-        {activeTab === 'advanced' && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-            <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex items-center gap-3">
-                <div className="p-2 bg-indigo-100 dark:bg-indigo-900/30 rounded-lg text-indigo-600 dark:text-indigo-400">
-                    <GitBranch size={20} />
-                </div>
-                <div>
-                    <h3 className="font-bold text-gray-900 dark:text-white">Template Registries</h3>
-                  <p className="text-xs text-gray-500 dark:text-gray-400">Connect Git repositories that supply ServiceBay templates and stacks.</p>
-                </div>
-                <div className="ml-auto">
-                    <label className="relative inline-flex items-center cursor-pointer">
-                        <input 
-                            type="checkbox" 
-                            className="sr-only peer" 
-                            checked={registriesEnabled}
-                      onChange={e => handleRegistriesToggle(e.target.checked)}
-                      disabled={saving}
-                        />
-                        <div className="w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-indigo-300 dark:peer-focus:ring-indigo-800 rounded-full peer dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-indigo-600"></div>
-                    </label>
-                </div>
-            </div>
-            
-            {registriesEnabled && (
-                <div className="p-6 space-y-6">
-                    <div className="grid grid-cols-1 md:grid-cols-12 gap-4 items-end">
-                        <div className="md:col-span-3">
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Name</label>
-                            <input 
-                                type="text" 
-                                value={newRegName}
-                                onChange={e => setNewRegName(e.target.value)}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-indigo-500 outline-none"
-                                placeholder="my-registry"
-                            />
-                        </div>
-                        <div className="md:col-span-5">
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Git URL</label>
-                            <input 
-                                type="text" 
-                                value={newRegUrl}
-                                onChange={e => setNewRegUrl(e.target.value)}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-indigo-500 outline-none"
-                                placeholder="https://github.com/user/repo.git"
-                            />
-                        </div>
-                        <div className="md:col-span-3">
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Branch (Optional)</label>
-                            <input 
-                                type="text" 
-                                value={newRegBranch}
-                                onChange={e => setNewRegBranch(e.target.value)}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-indigo-500 outline-none"
-                                placeholder="main"
-                            />
-                        </div>
-                        <div className="md:col-span-1">
-                            <button 
-                                onClick={handleAddRegistry}
-                              disabled={saving || !newRegName || !newRegUrl}
-                              className="w-full p-2 bg-indigo-600 text-white rounded-lg hover:bg-indigo-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex justify-center"
-                            >
-                                <Plus size={20} />
-                            </button>
-                        </div>
-                    </div>
-
-                    <div className="space-y-2">
-                        {registries.map(reg => (
-                            <div key={reg.name} className="flex items-center justify-between p-3 bg-gray-50 dark:bg-gray-900 rounded-lg border border-gray-200 dark:border-gray-700">
-                                <div className="flex items-center gap-4">
-                                    <div className="font-medium text-gray-900 dark:text-white">{reg.name}</div>
-                                    <div className="text-sm text-gray-500 dark:text-gray-400 font-mono">{reg.url}</div>
-                                    {reg.branch && (
-                                        <div className="text-xs px-2 py-1 bg-gray-200 dark:bg-gray-800 rounded text-gray-600 dark:text-gray-300 font-mono">
-                                            {reg.branch}
-                                        </div>
-                                    )}
-                                </div>
-                                <button 
-                                    onClick={() => handleRemoveRegistry(reg.name)}
-                                  disabled={saving}
-                                  className="text-gray-400 hover:text-red-500 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                                >
-                                    <Trash2 size={18} />
-                                </button>
-                            </div>
-                        ))}
-                        {registries.length === 0 && (
-                            <p className="text-sm text-gray-500 italic text-center py-4">No external registries configured. Only built-in templates are available.</p>
-                        )}
-                    </div>
-                </div>
-            )}
-        </div>
-        )}
-
-
-
-        {/* Email Notifications Section */}
-        {activeTab === 'notifications' && (
-        <div className="bg-white dark:bg-gray-800 rounded-xl border border-gray-200 dark:border-gray-700 shadow-sm overflow-hidden w-full">
-            <div className="p-4 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 flex items-center gap-3">
-                <div className="p-2 bg-blue-100 dark:bg-blue-900/30 rounded-lg text-blue-600 dark:text-blue-400">
-                    <Mail size={20} />
-                </div>
-                <div>
-                    <h3 className="font-bold text-gray-900 dark:text-white">Email Notifications</h3>
-                    <p className="text-xs text-gray-500 dark:text-gray-400">Configure SMTP settings for alerts</p>
-                </div>
-                <div className="ml-auto">
-                    <label className="relative inline-flex items-center cursor-pointer">
-                        <input 
-                            type="checkbox" 
-                            className="sr-only peer" 
-                            checked={emailEnabled}
-                      onChange={e => handleEmailEnabledToggle(e.target.checked)}
-                      disabled={saving}
-                        />
-                        <div className="w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-blue-300 dark:peer-focus:ring-blue-800 rounded-full peer dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-blue-600"></div>
-                    </label>
-                </div>
-            </div>
-            
-            {emailEnabled && (
-                <div className="p-6 space-y-6">
-                    <div className="bg-blue-50 dark:bg-blue-900/20 border border-blue-200 dark:border-blue-800 rounded-lg p-4 text-sm text-blue-800 dark:text-blue-200">
-                        <p className="font-medium mb-1">Need help finding these settings?</p>
-                        <ul className="list-disc list-inside space-y-1 opacity-90">
-                            <li><strong>Gmail:</strong> Host: <code>smtp.gmail.com</code>, Port: <code>587</code>. Use an <a href="https://myaccount.google.com/apppasswords" target="_blank" rel="noopener noreferrer" className="underline hover:text-blue-600">App Password</a> if 2FA is enabled.</li>
-                            <li><strong>Outlook:</strong> Host: <code>smtp.office365.com</code>, Port: <code>587</code>.</li>
-                            <li><strong>GMX:</strong> Host: <code>mail.gmx.net</code>, Port: <code>587</code>.</li>
-                        </ul>
-                    </div>
-
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                        <div>
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">SMTP Host</label>
-                            <input 
-                                type="text" 
-                                value={emailHost}
-                              onChange={e => setEmailHost(e.target.value)}
-                              onBlur={() => persistSettings()}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none"
-                                placeholder="smtp.gmail.com"
-                            />
-                        </div>
-                        <div>
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">SMTP Port</label>
-                            <input 
-                                type="number" 
-                                value={emailPort}
-                              onChange={e => setEmailPort(parseInt(e.target.value) || 0)}
-                              onBlur={() => persistSettings()}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none"
-                                placeholder="587"
-                            />
-                        </div>
-                        <div>
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Username</label>
-                            <input 
-                                type="text" 
-                                value={emailUser}
-                              onChange={e => setEmailUser(e.target.value)}
-                              onBlur={() => persistSettings()}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none"
-                                placeholder="user@example.com"
-                            />
-                        </div>
-                        <div>
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">Password</label>
-                            <input 
-                                type="password" 
-                                value={emailPass}
-                              onChange={e => setEmailPass(e.target.value)}
-                              onBlur={() => persistSettings()}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none"
-                                placeholder="••••••••"
-                            />
-                        </div>
-                        <div className="md:col-span-2">
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">From Address</label>
-                            <input 
-                                type="text" 
-                                value={emailFrom}
-                              onChange={e => setEmailFrom(e.target.value)}
-                              onBlur={() => persistSettings()}
-                              disabled={saving}
-                                className="w-full p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none"
-                                placeholder="ServiceBay <alerts@example.com>"
-                            />
-                        </div>
-                        <div className="md:col-span-2">
-                            <label className="flex items-center gap-2 cursor-pointer">
-                                <input 
-                                    type="checkbox" 
-                                    checked={emailSecure}
-                                onChange={e => handleEmailSecureToggle(e.target.checked)}
-                                disabled={saving}
-                                    className="w-4 h-4 text-blue-600 rounded border-gray-300 focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-800 focus:ring-2 dark:bg-gray-700 dark:border-gray-600"
-                                />
-                                <span className="text-sm text-gray-700 dark:text-gray-300">Use Secure Connection (TLS/SSL)</span>
-                            </label>
-                        </div>
-                    </div>
-
-                    <div className="border-t border-gray-200 dark:border-gray-700 pt-6">
-                        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">Recipients</label>
-                        <div className="flex gap-2 mb-3">
-                            <input 
-                                type="email" 
-                                value={newRecipient}
-                                onChange={e => setNewRecipient(e.target.value)}
-                                onKeyDown={e => e.key === 'Enter' && handleAddRecipient()}
-                              disabled={saving}
-                                className="flex-1 p-2 rounded-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 outline-none"
-                                placeholder="Add email address..."
-                            />
-                            <button 
-                                onClick={handleAddRecipient}
-                              disabled={saving || !newRecipient}
-                              className="p-2 bg-gray-100 dark:bg-gray-700 text-gray-600 dark:text-gray-300 rounded-lg hover:bg-gray-200 dark:hover:bg-gray-600 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-                            >
-                                <Plus size={20} />
-                            </button>
-                        </div>
-                        <div className="space-y-2">
-                            {emailRecipients.map(email => (
-                                <div key={email} className="flex items-center justify-between p-2 bg-gray-50 dark:bg-gray-900 rounded-lg border border-gray-200 dark:border-gray-700">
-                                    <span className="text-sm text-gray-700 dark:text-gray-300">{email}</span>
-                                    <button 
-                                        onClick={() => handleRemoveRecipient(email)}
-                                  disabled={saving}
-                                  className="text-gray-400 hover:text-red-500 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                                    >
-                                        <Trash2 size={16} />
-                                    </button>
-                                </div>
-                            ))}
-                            {emailRecipients.length === 0 && (
-                                <p className="text-sm text-gray-500 italic">No recipients added.</p>
-                            )}
-                        </div>
-                    </div>
-                </div>
-            )}
-        </div>
-        )}
-
-        {/* Log Level Control */}
-        {activeTab === 'advanced' && (
-        <LogLevelControl />
-        )}
-      </div>
-
-      {/* Update Progress Modal */}
-      {updateStatus !== 'idle' && (
-        <div className="fixed inset-0 bg-black/50 backdrop-blur-sm z-[70] flex items-center justify-center p-4">
-          <div className="bg-white dark:bg-gray-900 rounded-xl shadow-2xl max-w-md w-full border border-gray-200 dark:border-gray-800 overflow-hidden">
-            <div className="p-6 text-center space-y-6">
-              {updateStatus === 'error' ? (
-                <div className="w-16 h-16 bg-red-100 dark:bg-red-900/30 text-red-600 dark:text-red-400 rounded-full flex items-center justify-center mx-auto">
-                  <XCircle size={32} />
-                </div>
-              ) : updateStatus === 'restarting' ? (
-                <div className="w-16 h-16 bg-green-100 dark:bg-green-900/30 text-green-600 dark:text-green-400 rounded-full flex items-center justify-center mx-auto animate-pulse">
-                  <CheckCircle2 size={32} />
-                </div>
-              ) : (
-                <div className="w-16 h-16 bg-purple-100 dark:bg-purple-900/30 text-purple-600 dark:text-purple-400 rounded-full flex items-center justify-center mx-auto">
-                  <Loader2 size={32} className="animate-spin" />
-                </div>
-              )}
-              
-              <div>
-                <h3 className="text-xl font-bold text-gray-900 dark:text-white mb-2">
-                  {updateStatus === 'error' ? 'Update Failed' : 
-                   updateStatus === 'restarting' ? 'Update Complete' : 
-                   'Updating ServiceBay'}
-                </h3>
-                <p className="text-gray-500 dark:text-gray-400 text-sm">
-                  {updateStatus === 'error' ? updateError : updateMessage}
-                </p>
-              </div>
-
-              {updateStatus === 'updating' && (
-                <div className="w-full bg-gray-200 dark:bg-gray-700 rounded-full h-2.5 overflow-hidden">
-                  <div 
-                    className="bg-purple-600 h-2.5 rounded-full transition-all duration-300 ease-out" 
-                    style={{ width: `${updateProgress}%` }}
-                  ></div>
-                </div>
-              )}
-
-              {updateStatus === 'error' && (
-                <button
-                  onClick={() => setUpdateStatus('idle')}
-                  className="px-4 py-2 bg-gray-100 dark:bg-gray-800 text-gray-900 dark:text-white rounded-lg hover:bg-gray-200 dark:hover:bg-gray-700 transition-colors font-medium text-sm"
-                >
-                  Close
-                </button>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-
-      <ConfirmModal
-        isOpen={isUpdateModalOpen}
-        title="Update ServiceBay"
-        message={`Are you sure you want to update ServiceBay to version ${appUpdate?.latest?.version}? The service will restart automatically.`}
-        confirmText="Update Now"
-        onConfirm={confirmAppUpdate}
-        onCancel={() => setIsUpdateModalOpen(false)}
-      />
-
-      <ConfirmModal
-        isOpen={!!deleteTarget}
-        title="Delete Backup"
-        message={`Delete ${deleteTarget?.fileName || 'this backup'} permanently? This action cannot be undone.`}
-        confirmText={deletingBackup ? 'Deleting...' : 'Delete Backup'}
-        confirmDisabled={deletingBackup}
-        isDestructive
-        onConfirm={confirmDeleteBackup}
-        onCancel={() => !deletingBackup && setDeleteTarget(null)}
-      />
-
-      {restoreOverlayOpen && (
-        <div className="fixed inset-0 z-[90] flex items-stretch justify-end" onMouseDown={stopRestoreEvent} onClick={stopRestoreEvent}>
-          <div className="absolute inset-0 bg-gray-950/70 backdrop-blur-sm" onClick={handleRestoreBackdrop} />
-          <aside className="relative z-10 w-full max-w-3xl h-full bg-white dark:bg-gray-950 border-l border-gray-200 dark:border-gray-800 shadow-2xl flex flex-col">
-            <div className="flex items-center justify-between px-6 py-4 border-b border-gray-200 dark:border-gray-800">
-              <div>
-                <h3 className="text-lg font-semibold text-gray-900 dark:text-gray-100">Restore from Backup</h3>
-                <p className="text-xs text-gray-500 dark:text-gray-400">Select what to restore before applying changes.</p>
-              </div>
-              <button
-                type="button"
-                onClick={closeRestoreOverlay}
-                className="rounded-full p-2 text-gray-500 hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-100 hover:bg-gray-100 dark:hover:bg-gray-800"
-                aria-label="Close restore panel"
-              >
-                <X size={18} />
-              </button>
-            </div>
-
-            <div className="flex-1 overflow-y-auto px-6 py-5 space-y-6">
-              {!restorePreview ? (
-                <div
-                  className="border-2 border-dashed border-gray-300 dark:border-gray-700 rounded-xl p-8 text-center bg-gray-50 dark:bg-gray-900/40"
-                  onDrop={handleRestoreDrop}
-                  onDragOver={handleRestoreDragOver}
-                >
-                  <UploadCloud className="mx-auto text-gray-400" size={28} />
-                  <p className="mt-3 text-sm font-medium text-gray-700 dark:text-gray-200">Drop a backup archive here</p>
-                  <p className="text-xs text-gray-500 dark:text-gray-400">Supports .tar.gz exports from ServiceBay.</p>
-                  <div className="mt-4">
-                    <label
-                      htmlFor="restore-backup-file"
-                      className="inline-flex items-center gap-2 px-4 py-2 bg-white dark:bg-gray-900 border border-gray-300 dark:border-gray-700 rounded-lg text-sm text-gray-700 dark:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-800 cursor-pointer"
-                    >
-                      <UploadCloud size={16} />
-                      Select file
-                    </label>
-                    <input
-                      id="restore-backup-file"
-                      type="file"
-                      accept=".tar.gz"
-                      className="hidden"
-                      onChange={(event) => handleRestoreFromFile(event.target.files?.[0] || null)}
-                    />
-                  </div>
-                  {restoreUploadError && (
-                    <p className="mt-3 text-xs text-red-600 dark:text-red-400">{restoreUploadError}</p>
-                  )}
-                </div>
-              ) : restoreSelectionState ? (
-                <div className="space-y-3">
-                  {/* Source & Summary Bar */}
-                  <div className="rounded-lg border border-gray-200 dark:border-gray-800 p-4 bg-gray-50 dark:bg-gray-900/40 flex flex-wrap items-center justify-between gap-3">
-                    <div className="min-w-0">
-                      <p className="text-xs text-gray-500 dark:text-gray-400">Backup Source</p>
-                      <p className="text-sm font-mono text-gray-800 dark:text-gray-200 truncate">
-                        {restoreSource?.type === 'stored' ? restoreSource.fileName : 'Uploaded archive'}
-                      </p>
-                    </div>
-                    <div className="flex items-center gap-3">
-                      <span className="text-xs text-gray-500 dark:text-gray-400">{getRestoreSelectionSummary()}</span>
-                      <button
-                        type="button"
-                        onClick={() => {
-                          selectAllRestoreItems();
-                          void confirmRestoreBackup();
-                        }}
-                        disabled={restoringBackup}
-                        className="inline-flex items-center gap-2 px-3 py-1.5 text-xs font-medium rounded-lg border border-emerald-500 text-emerald-700 dark:text-emerald-300 dark:border-emerald-700 hover:bg-emerald-50 dark:hover:bg-emerald-900/20 transition-colors"
-                      >
-                        <RotateCcw size={14} />
-                        Restore all
-                      </button>
-                    </div>
-                  </div>
-
-                  {/* === Settings Section === */}
-                  <div className="rounded-lg border border-gray-200 dark:border-gray-800 overflow-hidden">
-                    <button
-                      type="button"
-                      onClick={() => toggleRestoreSection('settings')}
-                      className="w-full flex items-center gap-3 px-4 py-3 text-left hover:bg-gray-50 dark:hover:bg-gray-900/40 transition-colors"
-                    >
-                      {restoreExpandedSections.settings ? <ChevronDown size={16} className="text-gray-400 shrink-0" /> : <ChevronRight size={16} className="text-gray-400 shrink-0" />}
-                      <Settings size={16} className="text-gray-400 shrink-0" />
-                      <div className="flex-1 min-w-0">
-                        <span className="text-sm font-medium text-gray-900 dark:text-gray-100">Settings</span>
-                        <span className="ml-2 text-xs text-gray-500 dark:text-gray-400">
-                          {Object.values(restoreSelectionState.configFlags).filter(Boolean).length} of {Object.keys(restoreSelectionState.configFlags).length} selected
-                        </span>
-                      </div>
-                    </button>
-                    {restoreExpandedSections.settings && (
-                      <div className="px-4 pb-4 pt-1 border-t border-gray-100 dark:border-gray-800/50 grid gap-2.5">
-                        {([
-                          { key: 'externalLinks' as const, label: 'External links', summary: restorePreview.config.externalLinks.length === 0 ? 'None' : `${restorePreview.config.externalLinks.length} link${restorePreview.config.externalLinks.length !== 1 ? 's' : ''}` },
-                          { key: 'registries' as const, label: 'Registries', summary: restorePreview.config.registries.length === 0 ? 'None' : restorePreview.config.registries.map(r => r.name).join(', ') },
-                          { key: 'gateway' as const, label: 'Gateway', summary: restorePreview.config.gateway?.host || 'Not configured' },
-                          { key: 'notifications' as const, label: 'Notifications', summary: restorePreview.config.notifications ? `${restorePreview.config.notifications.host || 'SMTP'} → ${(restorePreview.config.notifications.to || []).join(', ') || 'no recipients'}` : 'Not configured' },
-                          { key: 'templateSettings' as const, label: 'Template settings', summary: restorePreview.config.templateSettings.length === 0 ? 'None' : `${restorePreview.config.templateSettings.length} key${restorePreview.config.templateSettings.length !== 1 ? 's' : ''}` },
-                          { key: 'logLevel' as const, label: 'Log level', summary: restorePreview.config.logLevel || 'default' },
-                          { key: 'update' as const, label: 'Auto-update', summary: restorePreview.config.update ? `${restorePreview.config.update.channel || 'stable'}${restorePreview.config.update.enabled === false ? ' (disabled)' : ''}` : 'Not configured' },
-                        ]).map(item => (
-                          <label key={item.key} className="flex items-center gap-3 text-sm text-gray-700 dark:text-gray-200 py-1">
-                            <input
-                              type="checkbox"
-                              className="rounded"
-                              checked={restoreSelectionState.configFlags[item.key]}
-                              onChange={() => toggleRestoreConfigFlag(item.key)}
-                            />
-                            <span className="font-medium min-w-[120px]">{item.label}</span>
-                            <span className="text-xs text-gray-500 dark:text-gray-400 truncate">{item.summary}</span>
-                          </label>
-                        ))}
-                      </div>
-                    )}
-                  </div>
-
-                  {/* === Nodes & Checks Section === */}
-                  <div className="rounded-lg border border-gray-200 dark:border-gray-800 overflow-hidden">
-                    <button
-                      type="button"
-                      onClick={() => toggleRestoreSection('infrastructure')}
-                      className="w-full flex items-center gap-3 px-4 py-3 text-left hover:bg-gray-50 dark:hover:bg-gray-900/40 transition-colors"
-                    >
-                      {restoreExpandedSections.infrastructure ? <ChevronDown size={16} className="text-gray-400 shrink-0" /> : <ChevronRight size={16} className="text-gray-400 shrink-0" />}
-                      <Activity size={16} className="text-gray-400 shrink-0" />
-                      <div className="flex-1 min-w-0">
-                        <span className="text-sm font-medium text-gray-900 dark:text-gray-100">Nodes & Monitoring</span>
-                        <span className="ml-2 text-xs text-gray-500 dark:text-gray-400">
-                          {Object.values(restoreSelectionState.nodes).filter(Boolean).length} node{Object.values(restoreSelectionState.nodes).filter(Boolean).length !== 1 ? 's' : ''},
-                          {' '}{Object.values(restoreSelectionState.checks).filter(Boolean).length} check{Object.values(restoreSelectionState.checks).filter(Boolean).length !== 1 ? 's' : ''}
-                        </span>
-                      </div>
-                    </button>
-                    {restoreExpandedSections.infrastructure && (
-                      <div className="px-4 pb-4 pt-1 border-t border-gray-100 dark:border-gray-800/50 space-y-4">
-                        {restorePreview.config.nodes.length > 0 && (
-                          <div>
-                            <p className="text-xs font-medium text-gray-500 dark:text-gray-400 uppercase tracking-wide mb-2">Nodes</p>
-                            <div className="grid gap-2">
-                              {restorePreview.config.nodes.map(node => (
-                                <label key={node.name} className="flex items-center gap-3 text-sm text-gray-700 dark:text-gray-200">
-                                  <input type="checkbox" className="rounded" checked={restoreSelectionState.nodes[node.name]} onChange={() => toggleRestoreNode(node.name)} />
-                                  <Server size={14} className="text-gray-400 shrink-0" />
-                                  <span className="font-medium">{node.name}</span>
-                                  <span className="text-xs text-gray-500 dark:text-gray-400">{node.uri}{node.default ? ' · Default' : ''}</span>
-                                </label>
-                              ))}
-                            </div>
-                          </div>
-                        )}
-                        {restorePreview.config.checks.length > 0 && (
-                          <div>
-                            <p className="text-xs font-medium text-gray-500 dark:text-gray-400 uppercase tracking-wide mb-2">Monitoring Checks</p>
-                            <div className="grid gap-2">
-                              {restorePreview.config.checks.map(check => (
-                                <label key={check.id} className="flex items-center gap-3 text-sm text-gray-700 dark:text-gray-200">
-                                  <input type="checkbox" className="rounded" checked={restoreSelectionState.checks[check.id]} onChange={() => toggleRestoreCheck(check.id)} />
-                                  <span className="font-medium">{check.name}</span>
-                                  {check.type && <span className="text-[10px] uppercase px-1.5 py-0.5 rounded bg-gray-100 dark:bg-gray-800 text-gray-500 dark:text-gray-400">{check.type}</span>}
-                                  {check.target && <span className="text-xs text-gray-500 dark:text-gray-400">{check.target}</span>}
-                                </label>
-                              ))}
-                            </div>
-                          </div>
-                        )}
-                        {restorePreview.config.nodes.length === 0 && restorePreview.config.checks.length === 0 && (
-                          <p className="text-xs text-gray-500 dark:text-gray-400 italic">No nodes or checks in this backup.</p>
-                        )}
-                      </div>
-                    )}
-                  </div>
-
-                  {/* === Systemd Files Section === */}
-                  {restorePreview.nodeFiles.length > 0 && (
-                    <div className="rounded-lg border border-gray-200 dark:border-gray-800 overflow-hidden">
-                      <button
-                        type="button"
-                        onClick={() => toggleRestoreSection('files')}
-                        className="w-full flex items-center gap-3 px-4 py-3 text-left hover:bg-gray-50 dark:hover:bg-gray-900/40 transition-colors"
-                      >
-                        {restoreExpandedSections.files ? <ChevronDown size={16} className="text-gray-400 shrink-0" /> : <ChevronRight size={16} className="text-gray-400 shrink-0" />}
-                        <FolderOpen size={16} className="text-gray-400 shrink-0" />
-                        <div className="flex-1 min-w-0">
-                          <span className="text-sm font-medium text-gray-900 dark:text-gray-100">Systemd Files</span>
-                          <span className="ml-2 text-xs text-gray-500 dark:text-gray-400">
-                            {Object.values(restoreSelectionState.nodeFiles).reduce((sum, files) => sum + Object.values(files).filter(Boolean).length, 0)} of {restorePreview.nodeFiles.reduce((sum, g) => sum + g.files.length, 0)} files selected
-                          </span>
-                        </div>
-                      </button>
-                      {restoreExpandedSections.files && (
-                        <div className="border-t border-gray-100 dark:border-gray-800/50">
-                          {restorePreview.nodeFiles.map(group => {
-                            const selectedCount = Object.values(restoreSelectionState.nodeFiles[group.nodeName] || {}).filter(Boolean).length;
-                            const allSelected = selectedCount === group.files.length;
-                            const serviceGroups = groupFilesByService(group.files);
-                            return (
-                              <div key={group.nodeName} className="border-b border-gray-100 dark:border-gray-800/50 last:border-b-0">
-                                <div className="flex items-center justify-between gap-2 px-4 py-2.5 bg-gray-50/50 dark:bg-gray-900/20">
-                                  <div className="flex items-center gap-3">
-                                    <input
-                                      type="checkbox"
-                                      className="rounded"
-                                      checked={allSelected}
-                                      onChange={() => toggleAllNodeFiles(group.nodeName, !allSelected)}
-                                    />
-                                    <span className="text-sm font-medium text-gray-900 dark:text-gray-100">{group.nodeName}</span>
-                                    <span className="text-xs text-gray-500 dark:text-gray-400">{selectedCount}/{group.files.length} files</span>
-                                  </div>
-                                  <div className="flex items-center gap-2 text-xs text-gray-500 dark:text-gray-400">
-                                    <span>Target:</span>
-                                    <select
-                                      value={restoreSelectionState.targetNodes[group.nodeName]}
-                                      onChange={(event) => updateRestoreTargetNode(group.nodeName, event.target.value)}
-                                      className="bg-white dark:bg-gray-900 border border-gray-300 dark:border-gray-700 text-gray-700 dark:text-gray-200 rounded px-2 py-1 text-xs"
-                                    >
-                                      {availableRestoreTargets.map(target => (
-                                        <option key={target} value={target}>{target}</option>
-                                      ))}
-                                    </select>
-                                  </div>
-                                </div>
-                                <div className="max-h-80 overflow-y-auto">
-                                  {serviceGroups.map(sg => {
-                                    const sgSelectedCount = sg.files.filter(f => restoreSelectionState.nodeFiles[group.nodeName]?.[f.relativePath]).length;
-                                    const sgAllSelected = sgSelectedCount === sg.files.length;
-                                    const sgKey = `files-${group.nodeName}-${sg.service}`;
-                                    const sgExpanded = restoreExpandedSections[sgKey];
-                                    const displayName = sg.service === '_other' ? 'Other files' : sg.service;
-                                    return (
-                                      <div key={sg.service} className="border-b border-gray-50 dark:border-gray-800/30 last:border-b-0">
-                                        <div className="flex items-center gap-2 px-4 py-2 hover:bg-gray-50 dark:hover:bg-gray-900/30">
-                                          <input
-                                            type="checkbox"
-                                            className="rounded"
-                                            checked={sgAllSelected}
-                                            onChange={() => toggleServiceGroupFiles(group.nodeName, sg.files, !sgAllSelected)}
-                                          />
-                                          <button
-                                            type="button"
-                                            onClick={() => toggleRestoreSection(sgKey)}
-                                            className="flex items-center gap-1.5 flex-1 min-w-0 text-left"
-                                          >
-                                            {sgExpanded ? <ChevronDown size={12} className="text-gray-400 shrink-0" /> : <ChevronRight size={12} className="text-gray-400 shrink-0" />}
-                                            <span className="text-xs font-medium text-gray-800 dark:text-gray-200 capitalize">{displayName}</span>
-                                            <span className="text-[10px] text-gray-400 dark:text-gray-500">{sgSelectedCount}/{sg.files.length}</span>
-                                          </button>
-                                        </div>
-                                        {sgExpanded && (
-                                          <div className="pl-10 pr-4 pb-1">
-                                            {sg.files.map(file => (
-                                              <div key={file.relativePath} className="flex items-center gap-3 py-1 text-xs text-gray-600 dark:text-gray-300">
-                                                <input
-                                                  type="checkbox"
-                                                  className="rounded"
-                                                  checked={Boolean(restoreSelectionState.nodeFiles[group.nodeName]?.[file.relativePath])}
-                                                  onChange={() => toggleRestoreFile(group.nodeName, file.relativePath)}
-                                                />
-                                                <span className="flex-1 font-mono truncate">{file.fileName}</span>
-                                                <button
-                                                  type="button"
-                                                  onClick={() => handleRestoreFilePreview(group.nodeName, file.relativePath)}
-                                                  className="shrink-0 p-1 rounded text-gray-400 hover:text-gray-600 dark:hover:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-800"
-                                                  title="Preview file"
-                                                >
-                                                  <Eye size={14} />
-                                                </button>
-                                              </div>
-                                            ))}
-                                          </div>
-                                        )}
-                                      </div>
-                                    );
-                                  })}
-                                </div>
-                              </div>
-                            );
-                          })}
-                        </div>
-                      )}
-                    </div>
-                  )}
-
-                  {/* === Service Data Section === */}
-                  {restorePreview.serviceData && restorePreview.serviceData.length > 0 && (
-                    <div className="rounded-lg border border-gray-200 dark:border-gray-800 overflow-hidden">
-                      <button
-                        type="button"
-                        onClick={() => toggleRestoreSection('serviceData')}
-                        className="w-full flex items-center gap-3 px-4 py-3 text-left hover:bg-gray-50 dark:hover:bg-gray-900/40 transition-colors"
-                      >
-                        {restoreExpandedSections.serviceData ? <ChevronDown size={16} className="text-gray-400 shrink-0" /> : <ChevronRight size={16} className="text-gray-400 shrink-0" />}
-                        <Database size={16} className="text-gray-400 shrink-0" />
-                        <div className="flex-1 min-w-0">
-                          <span className="text-sm font-medium text-gray-900 dark:text-gray-100">Service Data</span>
-                          <span className="ml-2 text-xs text-gray-500 dark:text-gray-400">
-                            {Object.values(restoreSelectionState.serviceData).reduce((sum, fm) => sum + Object.values(fm).filter(Boolean).length, 0)} file{Object.values(restoreSelectionState.serviceData).reduce((sum, fm) => sum + Object.values(fm).filter(Boolean).length, 0) !== 1 ? 's' : ''} selected
-                          </span>
-                        </div>
-                      </button>
-                      {restoreExpandedSections.serviceData && (
-                        <div className="px-4 pb-4 pt-1 border-t border-gray-100 dark:border-gray-800/50 space-y-3">
-                          {restorePreview.serviceData.map(sd => {
-                            const label = sd.name.replace(/-/g, '/').replace(/^\//, '');
-                            const fileCategories = groupServiceDataFiles(sd.files);
-                            const sdKey = `sd-${sd.name}`;
-                            const sdExpanded = restoreExpandedSections[sdKey];
-                            const selectedCount = restoreSelectionState.serviceData[sd.name]
-                              ? Object.values(restoreSelectionState.serviceData[sd.name]).filter(Boolean).length
-                              : 0;
-                            const allSelected = selectedCount === sd.files.length;
-
-                            return (
-                              <div key={sd.name} className="rounded border border-gray-100 dark:border-gray-800/50">
-                                <div className="flex items-center gap-2 px-3 py-2">
-                                  <input
-                                    type="checkbox"
-                                    className="rounded"
-                                    checked={allSelected}
-                                    onChange={() => setRestoreSelectionState(prev => {
-                                      if (!prev) return prev;
-                                      const newVal = !allSelected;
-                                      return {
-                                        ...prev,
-                                        serviceData: {
-                                          ...prev.serviceData,
-                                          [sd.name]: Object.fromEntries(sd.files.map(f => [f, newVal]))
-                                        }
-                                      };
-                                    })}
-                                  />
-                                  <button type="button" onClick={() => toggleRestoreSection(sdKey)} className="flex items-center gap-2 flex-1 min-w-0">
-                                    {sdExpanded ? <ChevronDown size={14} className="text-gray-400 shrink-0" /> : <ChevronRight size={14} className="text-gray-400 shrink-0" />}
-                                    <HardDrive size={14} className="text-gray-400 shrink-0" />
-                                    <div className="flex flex-col items-start min-w-0">
-                                      <div className="flex items-center gap-2">
-                                        <span className="text-sm font-medium text-gray-900 dark:text-gray-100">{label}</span>
-                                        <span className="text-xs text-gray-500 dark:text-gray-400">{selectedCount}/{sd.files.length}</span>
-                                      </div>
-                                      {(sd.sourcePath || sd.nodeName) && (
-                                        <span className="text-[11px] text-gray-400 dark:text-gray-500 font-mono truncate max-w-full">
-                                          → {sd.nodeName ? `${sd.nodeName}:` : ''}{sd.sourcePath}
-                                        </span>
-                                      )}
-                                    </div>
-                                  </button>
-                                  {/* Quick-select buttons for categories */}
-                                  <div className="flex items-center gap-1 shrink-0">
-                                    {fileCategories.map(cat => (
-                                      <button
-                                        key={cat.category}
-                                        type="button"
-                                        title={`Select only ${SERVICE_DATA_CATEGORY_LABELS[cat.category].toLowerCase()} (${cat.files.length} files)`}
-                                        onClick={() => setRestoreSelectionState(prev => {
-                                          if (!prev) return prev;
-                                          // Select only files in this category, deselect everything else
-                                          const newFiles: Record<string, boolean> = {};
-                                          for (const f of sd.files) newFiles[f] = false;
-                                          for (const f of cat.files) newFiles[f] = true;
-                                          return {
-                                            ...prev,
-                                            serviceData: { ...prev.serviceData, [sd.name]: newFiles }
-                                          };
-                                        })}
-                                        className="px-1.5 py-0.5 text-xs rounded border border-gray-200 dark:border-gray-700 hover:bg-gray-100 dark:hover:bg-gray-800 transition-colors"
-                                      >
-                                        {SERVICE_DATA_CATEGORY_ICONS[cat.category]}
-                                      </button>
-                                    ))}
-                                  </div>
-                                </div>
-
-                                {sdExpanded && (
-                                  <div className="px-3 pb-3 pt-1 border-t border-gray-100 dark:border-gray-800/50 space-y-2">
-                                    {fileCategories.map(cat => {
-                                      const catKey = `sd-${sd.name}-${cat.category}`;
-                                      const catExpanded = restoreExpandedSections[catKey];
-                                      const catSelectedCount = cat.files.filter(f => restoreSelectionState.serviceData[sd.name]?.[f]).length;
-                                      const catAllSelected = catSelectedCount === cat.files.length;
-
-                                      return (
-                                        <div key={cat.category}>
-                                          <div className="flex items-center gap-2">
-                                            <input
-                                              type="checkbox"
-                                              className="rounded"
-                                              checked={catAllSelected}
-                                              onChange={() => setRestoreSelectionState(prev => {
-                                                if (!prev) return prev;
-                                                const updated = { ...prev.serviceData[sd.name] };
-                                                const newVal = !catAllSelected;
-                                                for (const f of cat.files) updated[f] = newVal;
-                                                return {
-                                                  ...prev,
-                                                  serviceData: { ...prev.serviceData, [sd.name]: updated }
-                                                };
-                                              })}
-                                            />
-                                            <button type="button" onClick={() => toggleRestoreSection(catKey)} className="flex items-center gap-1.5 flex-1 min-w-0">
-                                              {catExpanded ? <ChevronDown size={12} className="text-gray-400 shrink-0" /> : <ChevronRight size={12} className="text-gray-400 shrink-0" />}
-                                              <span className="text-xs">{SERVICE_DATA_CATEGORY_ICONS[cat.category]}</span>
-                                              <span className="text-xs font-medium text-gray-700 dark:text-gray-300">{SERVICE_DATA_CATEGORY_LABELS[cat.category]}</span>
-                                              <span className="text-xs text-gray-500 dark:text-gray-400">{catSelectedCount}/{cat.files.length}</span>
-                                            </button>
-                                          </div>
-                                          {catExpanded && (
-                                            <div className="ml-6 mt-1 space-y-0.5">
-                                              {cat.files.map(file => (
-                                                <label key={file} className="flex items-center gap-2 text-xs text-gray-600 dark:text-gray-300 py-0.5">
-                                                  <input
-                                                    type="checkbox"
-                                                    className="rounded"
-                                                    checked={Boolean(restoreSelectionState.serviceData[sd.name]?.[file])}
-                                                    onChange={() => setRestoreSelectionState(prev => {
-                                                      if (!prev) return prev;
-                                                      const updated = { ...prev.serviceData[sd.name] };
-                                                      updated[file] = !updated[file];
-                                                      return {
-                                                        ...prev,
-                                                        serviceData: { ...prev.serviceData, [sd.name]: updated }
-                                                      };
-                                                    })}
-                                                  />
-                                                  <span className="font-mono truncate">{file}</span>
-                                                </label>
-                                              ))}
-                                            </div>
-                                          )}
-                                        </div>
-                                      );
-                                    })}
-                                  </div>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      )}
-                    </div>
-                  )}
-                </div>
-              ) : null}
-            </div>
-
-            {restorePreview && restoreSelectionState && (
-              <div className="px-6 py-4 border-t border-gray-200 dark:border-gray-800 flex items-center justify-between">
-                <p className="text-xs text-gray-500 dark:text-gray-400">{getRestoreSelectionSummary()}</p>
-                <button
-                  onClick={confirmRestoreBackup}
-                  disabled={restoringBackup}
-                  className="inline-flex items-center gap-2 px-4 py-2 bg-emerald-600 text-white text-sm rounded-lg shadow-sm hover:bg-emerald-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  {restoringBackup ? <Loader2 className="animate-spin" size={16} /> : <RotateCcw size={16} />}
-                  {restoringBackup ? 'Restoring...' : 'Restore Selected'}
-                </button>
-              </div>
-            )}
-          </aside>
-        </div>
-      )}
-
-      {restoreFilePreview && (
-        <div className="fixed inset-0 z-[100] flex items-center justify-center" onMouseDown={stopRestoreEvent} onClick={stopRestoreEvent}>
-          <div className="absolute inset-0 bg-gray-950/70 backdrop-blur-sm" onClick={() => setRestoreFilePreview(null)} />
-          <div className="relative z-10 w-full max-w-5xl max-h-[85vh] bg-white dark:bg-gray-950 border border-gray-200 dark:border-gray-800 rounded-xl shadow-2xl overflow-hidden flex flex-col">
-            <div className="flex items-center justify-between px-5 py-3 border-b border-gray-200 dark:border-gray-800">
-              <div>
-                <p className="text-sm font-semibold text-gray-900 dark:text-gray-100">Backup File Preview</p>
-                <p className="text-xs text-gray-500 dark:text-gray-400">
-                  {restoreFilePreview.nodeName} · <span className="font-mono">{restoreFilePreview.relativePath}</span>
-                </p>
-              </div>
-              <button
-                type="button"
-                onClick={() => setRestoreFilePreview(null)}
-                className="rounded-full p-2 text-gray-500 hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-100 hover:bg-gray-100 dark:hover:bg-gray-800"
-                aria-label="Close file preview"
-              >
-                <X size={18} />
-              </button>
-            </div>
-            <div className="flex-1 overflow-y-auto bg-gray-50 dark:bg-gray-900/40 p-4">
-              {restoreFilePreview.loading ? (
-                <div className="flex items-center gap-2 text-sm text-gray-500 dark:text-gray-400">
-                  <Loader2 size={16} className="animate-spin" />
-                  Loading file...
-                </div>
-              ) : restoreFilePreviewError ? (
-                <p className="text-sm text-red-600 dark:text-red-400">{restoreFilePreviewError}</p>
-              ) : (
-                <FileViewer
-                  content={restoreFilePreview.content}
-                  language={resolveFilePreviewLanguage(restoreFilePreview.relativePath)}
-                />
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-
-      <SSHSetupModal 
-        isOpen={isSSHModalOpen}
-        onClose={() => setIsSSHModalOpen(false)}
-        initialHost={sshModalDefaults.host}
-        initialPort={sshModalDefaults.port}
-        initialUser={sshModalDefaults.user}
-      />
-    </div>
-  );
+export default function SettingsRootPage() {
+  redirect('/settings/nodes');
 }
